### 2025-09-15 #8999 @editorialbot
[PRE REVIEW]: topoptlab: An Open and Modular Framework for Benchmarking and Research in Topology Optimization
**Submitting author:** @stefanhiemer (<a href="http://orcid.org/0000-0001-6562-4883">Stefan Hiemer</a>)
**Repository:** https://github.com/stefanhiemer/topoptlab
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.2
