### 2025-09-15 a3f58b9
added DOIs to paper.bib.
