+++ paper/paper.bib
@@ -1,9 +1,14 @@
-@online{GitRepo,
-  author = {Stefan Hiemer},
-  title = {Topoptlab GitHub repository},
-  year = 2025,
-  url = {https://github.com/stefanhiemer/topoptlab},
-  urldate = {2025-08-25}
+@software{GitRepo,
+  author       = {stefanhiemer},
+  title        = {stefanhiemer/topoptlab: Release version 0.0.2 for
+                   JOSS submission
+                  },
+  month        = sep,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {0.0.2},
+  doi          = {10.5281/zenodo.17118124},
+  url          = {https://doi.org/10.5281/zenodo.17118124},
 }
 
 @article{sigmund200199,
@@ -14,7 +19,8 @@
   number={2},
   pages={120--127},
   year={2001},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s001580050176}
 }
 
 @article{andreassen2011efficient,
@@ -25,7 +31,8 @@
   number={1},
   pages={1--16},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00158-010-0594-7}
 }
 
 @article{ferrari2020new,
@@ -36,7 +43,8 @@
   number={4},
   pages={2211--2228},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00158-020-02629-w}
 }
 
 @article{wang2021comprehensive,
@@ -47,7 +55,8 @@
   number={5},
   pages={2827--2880},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00158-021-03050-7}
 }
 
 #
@@ -58,7 +67,8 @@
   volume={19},
   pages={101--113},
   year={2013},
-  publisher={Forschungszentrum J{\"u}lich J{\"u}lich, Germany}
+  publisher={Forschungszentrum J{\"u}lich J{\"u}lich, Germany},
+  note         = {No DOI found}
 }
 
 @article{arndt2021deal,
@@ -68,7 +78,8 @@
   volume={81},
   pages={407--422},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.camwa.2020.02.022}
 }
 
 @article{alnaes2014unified,
@@ -79,7 +90,8 @@
   number={2},
   pages={1--37},
   year={2014},
-  publisher={ACM New York, NY, USA}
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/2566630}
 }
 
 @article{scroggs2022construction,
@@ -90,7 +102,8 @@
   number={2},
   pages={1--23},
   year={2022},
-  publisher={ACM New York, NY}
+  publisher={ACM New York, NY},
+  doi={10.1145/3524456}
 }
 
 @article{scroggs2022basix,
@@ -100,13 +113,15 @@
   volume={7},
   number={73},
   pages={3982},
-  year={2022}
+  year={2022},
+  doi={10.21105/joss.03982}
 }
 
 @article{baratta2023dolfinx,
   title={DOLFINx: the next generation FEniCS problem solving environment},
   author={Baratta, Igor A and Dean, Joseph P and Dokken, J{\o}rgen S and Habera, Michal and HALE, Jack and Richardson, Chris N and Rognes, Marie E and Scroggs, Matthew W and Sime, Nathan and Wells, Garth N},
-  year={2023}
+  year={2023},
+  doi={10.5281/zenodo.10447666}
 }
 
 #
@@ -118,7 +133,8 @@
   volume={6},
   number={64},
   pages={3556},
-  year={2021}
[…]
