### 2025-09-15 #8999 @danielskatz
editorialbot commands need to be the first entry in a new comment.
