+++ docs/source/references.bib
@@ -1895,3 +1895,16 @@ jorunal={ }
   year={2008},
   publisher={Springer Science \& Business Media}
 }
+
+# multigrid in TO
+
+@article{amir2014multigrid,
+  title={On multigrid-CG for efficient topology optimization},
+  author={Amir, Oded and Aage, Niels and Lazarov, Boyan S},
+  journal={Structural and Multidisciplinary Optimization},
+  volume={49},
+  number={5},
+  pages={815--829},
+  year={2014},
+  publisher={Springer}
+}
