+++ docs/source/introduction/references.md
@@ -1,0 +1,4 @@
+# References
+
+```{bibliography}
+:style: unsrt
