### 2025-09-18 a7a29b9
Fix missing transpose in docu.
