### 2025-09-23 6903bc3
small refactoring in symfem to code conversion.
