### 2025-09-24 e15e7a4
started towards nonlinear elasticity with symfem, but still bug there. also small fixes.
