+++ topoptlab/symbolic/lin_elastic.py
@@ -48,7 +48,13 @@ def small_strain_matrix(ndim: int, nd_inds: list,
                         basis: List,
                         isoparam_kws: Dict) -> MatrixFunction:
     """
-    Create the small strain matrix commonly referred to as B matrix.
+    Create the small strain matrix to calculate engineering/infinitesimal 
+    strains via
+    
+    eps = B @ u
+    
+    where u are nodal displacements and eps_e a column vector of 
+    length (ndim**2 + ndim)/2.
 
     Parameters
     ----------
@@ -64,7 +70,7 @@ def small_strain_matrix(ndim: int, nd_inds: list,
     Returns
     -------
     bmatrix : symfem.functions.Matrixfunction
-        small displacement matrix.
+        small displacement matrix shape ( (ndim**2+ndim)/2 , n_nodes*ndim ).
     """
 
     nrows = int((ndim**2 + ndim) /2)
@@ -88,4 +94,49 @@ def small_strain_matrix(ndim: int, nd_inds: list,
         #
         i,j = (i+1)%ndim , (j+1)%ndim
     return MatrixFunction(bmatrix)
+
+def dispgrad_matrix(ndim: int, nd_inds: list,
+                    basis: List,
+                    isoparam_kws: Dict) -> MatrixFunction:
+    """
+    Create the matrix to calculate the flattened displacement gradient via
+    
+    h = B @ u
+    
+    where u are nodal displacements and h a column vector of length ndim**2. 
+    You recover the full deformation gradient by 
+    
+    H = h.reshape(ndim,ndim).
+
+    Parameters
+    ----------
+    ndim : int
+        number of spatial dimensions.
+    nd_inds : list
+        node indices.
+    basis : list
+        list of basis functions as generated by base_cell.
+    isoparam_kws : dictionary
+        keywords for the isoparametric mapping.
+
+    Returns
+    -------
+    dispgrad_matrix : symfem.functions.Matrixfunction
+        matrix to calculate flattened displacement gradient.
+    """
+    #
+    nrows = int(ndim**2)
+    ncols = int(ndim * len(nd_inds))
+    # compute gradients of basis functions
+    Jinv = jacobian(ndim=ndim,
+                    return_J=False, return_inv=True, return_det=False,
+                    **isoparam_kws)
+    gradN_T = (VectorFunction(basis).grad(ndim)@Jinv.transpose()).transpose()
+    #
+    dispgrad_matrix = [[0 for j in range(ncols)] for i in range(nrows)]
+    # 
+    for i in range(ndim):
+        for j in range(ndim): 
+            dispgrad_matrix[i*ndim + j][i::ndim] = gradN_T[j]
+    return MatrixFunction(dispgrad_matrix)
 
