### 2025-09-25 5a7d284
Symbolic strain and deformation measures work now. Still need to write tests and symbolic eigenvalues and eigenvectors for dim < 4 needed.
