+++ examples/symfem/strain_measures.py
@@ -1,16 +1,16 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
-from sympy import simplify
+from sympy import expand,simplify,Expr
 from symfem.functions import MatrixFunction
-
 
 from topoptlab.symbolic.cell import base_cell
 from topoptlab.symbolic.matrix_utils import simplify_matrix, generate_constMatrix
+from topoptlab.symbolic.matrix_utils import eye,to_square,inverse
 from topoptlab.symbolic.lin_elastic import small_strain_matrix,dispgrad_matrix
 from topoptlab.symbolic.voigt import convert_from_voigt
 
-def engstrain(ndim,
-              element_type="Lagrange",
-              order=1):
+def eng_strain(ndim: int,
+               element_type: str= "Lagrange",
+               order: int = 1) -> MatrixFunction:
     """
     Symbolically compute engineering strain.
 
@@ -25,13 +25,12 @@ def engstrain(ndim,
 
     Returns
     -------
-    nodal_forces : list
-        symbolic stiffness matrix as list of lists .
+    eng_strains : MatrixFunction
+        engineering strains in Voigt notation shape ( (ndim**2 + ndim)/2 ) .
 
     """
     #
     vertices, nd_inds, ref, basis  = base_cell(ndim)
-    print(basis)
     # small strain matrix
     b = small_strain_matrix(ndim=ndim,
                             nd_inds=nd_inds,
@@ -42,11 +41,13 @@ def engstrain(ndim,
     u = generate_constMatrix(1,b.shape[1], "u")
     return simplify_matrix( b@u )
 
-def dispgrad(ndim,
-             element_type="Lagrange",
-             order=1):
+def disp_grad(ndim: int,
+              element_type: str = "Lagrange",
+              order: int = 1) -> MatrixFunction:
     """
-    Symbolically compute displacement gradient.
+    Symbolically compute displacement gradient:
+        
+    H_ij = d u_i / d x_j
 
     Parameters
     ----------
@@ -59,13 +60,12 @@ def dispgrad(ndim,
 
     Returns
     -------
-    nodal_forces : list
-        symbolic stiffness matrix as list of lists .
+    H : symfem.MatrixFunction
+        symbolic displacement gradient of shape (ndim,ndim) .
 
     """
     #
     vertices, nd_inds, ref, basis  = base_cell(ndim)
-    print(basis)
     # small strain matrix
     b = dispgrad_matrix(ndim=ndim,
                         nd_inds=nd_inds,
@@ -74,13 +74,15 @@ def dispgrad(ndim,
                                       "order": order})
     # nodal displacements
     u = generate_constMatrix(1,b.shape[1], "u")
-    return convert_from_voigt(simplify_matrix( b@u ))
+    return to_square(simplify_matrix( b@u ),order="C")
 
-def defgrad(ndim,
-            element_type="Lagrange",
-            order=1):
+def def_grad(ndim: int,
+             element_type: str = "Lagrange",
+             order: int = 1):
     """
     Symbolically compute deformation gradient.
+    
+    F_ij = delta_ij + d u_i / d x_j
 
     Parameters
     ----------
@@ -93,41 +95,148 @@ def defgrad(ndim,
 
     Returns
     -------
-    nodal_forces : list
-        symbolic stiffness matrix as list of lists .
+    F : symfem.MatrixFunction
+        symbolic deformation gradient of shape (ndim,ndim) .
 
     """
     #
-    eye = MatrixFunction([[0 if i!=j else 1 for j in range(ndim)] \
-                          for i in range(ndim)])
+    I = eye(size=ndim)
     #
-    H = dispgrad(ndim,
-                 element_type="Lagrange",
-                 order=1)
-    return eye + H
[…]
