+++ topoptlab/symbolic/matrix_utils.py
@@ -1,9 +1,10 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 from typing import List, Tuple, Union
 from itertools import product
+from math import sqrt
 
 from sympy import symbols, Symbol
-from symfem.functions import MatrixFunction
+from symfem.functions import ScalarFunction,MatrixFunction
 
 def generate_constMatrix(ncol: int, nrow: int, name: str,
                          symmetric: bool = False,
@@ -96,3 +97,112 @@ def simplify_matrix(M: Union[List,MatrixFunction]) -> MatrixFunction:
     for i,j in product(range(M.shape[0]),range(M.shape[1])):
         M_new[i][j] = M[i,j].as_sympy().simplify()
     return MatrixFunction(M_new)
+
+def eye(size: int) -> MatrixFunction:
+    """
+    Return identity matrix of given size.
+
+    Parameters
+    ----------
+    size : int
+        size of identity matrix.
+
+    Returns
+    -------
+    eye : symfem.functions.MatrixFunction
+        identity matrix of shape (size,size).
+    """
+    
+    return MatrixFunction([[0 if i!=j else 1 for j in range(size)] \
+                          for i in range(size)])
+
+def diag(vals: List) -> MatrixFunction:
+    """
+    Return diagonal matrix with given diagonal values.
+
+    Parameters
+    ----------
+    vals : list
+        values of diagonal.
+
+    Returns
+    -------
+    diag : symfem.functions.MatrixFunction
+        diagonal matrix.
+    """
+    size=len(vals)
+    return MatrixFunction([[0 if i!=j else vals[i] for j in range(size)] \
+                          for i in range(size)])
+
+def to_square(v, order: str = "C") -> MatrixFunction:
+    """
+    Return MatrixFunction of shape (n**2,1) reshape (n,n) either in 'C' or 'F'
+    order analogously to numpy reshaping.
+
+    Parameters
+    ----------
+    v : symfem.functions.MatrixFunction
+        matrix of shape (n**2,1).
+    order : str
+        order of reshaping
+
+    Returns
+    -------
+    M : symfem.functions.MatrixFunction
+        matrix reshaped to shape (n,n)
+    """
+    n = int( sqrt(v.shape[0]) )
+    M = []
+    if order == "F":
+        for i in range(n):
+            M.append([v[j*n + i,0] for j in range(n)])
+    elif order == "C":
+        for i in range(n):
+            M.append([v[i*n+j,0] for j in range(n)])
+    return MatrixFunction(M)
+
+def inverse(A: MatrixFunction,
+            Adet: Union[None, ScalarFunction] = None) -> MatrixFunction:
+    """
+    Return the symbolic inverse of a square matrix using explicit adjugate 
+    formulas. Does not work if dim(A) > 3.
+
+    Parameters
+    ----------
+    A : symfem.functions.MatrixFunction
+       square matrix of shape (ndim, ndim).
+   Adet : symfem.functions.ScalarFunction or None, optional
+       determinant of ``A``. Providing it allows reuse.
+    Returns
+    -------
+    Ainv : symfem.functions.MatrixFunction
+       Matrix inverse of shape (ndim, ndim).
+
+    """
+    #
+    ndim = A.shape[0]
+    if ndim > 3:
[…]
