### 2025-09-30 8757c55
Added some new examples cases for 3D.
