+++ tests/test_linsolver.py
@@ -1,14 +1,18 @@
+from functools import partial
+
 from numpy.testing import assert_almost_equal
 from numpy import array
-from scipy.sparse import csc_matrix
+from scipy.sparse import csc_matrix,eye
 import pytest
 
-from topoptlab.linear_solvers import gauss_seidel ,smoothed_jacobi, modified_richardson, successive_overrelaxation
+from topoptlab.linear_solvers import gauss_seidel ,smoothed_jacobi,\
+    modified_richardson, successive_overrelaxation,pcg
 @pytest.mark.parametrize('fun',
                          [(gauss_seidel),
                           (smoothed_jacobi),
                           (modified_richardson),
-                          (successive_overrelaxation)])
+                          (successive_overrelaxation),
+                          ( partial(pcg,P=eye(3,3)) ) ])
 
 def test_linsystem(fun):
     
