### 2025-10-01 27a8da3
Demonstration case for using topoptlab as library plus some small refactoring.
