+++ examples/topology_optimization/compliance_minimization/TO_from_scratch.py
@@ -1,0 +1,198 @@
+# A 165 LINE TOPOLOGY OPTIMIZATION CODE BY NIELS AAGE AND VILLADS EGEDE JOHANSEN, JANUARY 2013
+# rewrite with the topoptlab package by Stefan Hiemer (January 2025)
+from typing import Callable, Union
+
+import numpy as np
+from scipy.sparse import coo_matrix
+from scipy.sparse.linalg import spsolve
+
+from matplotlib import colors
+import matplotlib.pyplot as plt
+
+from topoptlab.filter.matrix_filter import assemble_matrix_filter
+from topoptlab.fem import create_matrixinds,assemble_matrix,apply_bc
+from topoptlab.example_bc.lin_elast import mbb_2d,mbb_3d
+from topoptlab.elements.bilinear_quadrilateral import create_edofMat as create_edofMat2d
+from topoptlab.elements.trilinear_hexahedron import create_edofMat as create_edofMat3d
+from topoptlab.elements.linear_elasticity_2d import lk_linear_elast_2d
+from topoptlab.elements.linear_elasticity_3d import lk_linear_elast_3d
+from topoptlab.optimizer.optimality_criterion import oc_top88
+from topoptlab.solve_linsystem import solve_lin
+# MAIN DRIVER
+def main(nelx: int, nely: int, nelz: Union[None,int],
+         volfrac: float,
+         penal: float,
+         rmin: float, ft: int,
+         bcs: Callable,
+         lk: Callable):
+    """
+    Topology optimization for maximum stiffness with the SIMP method based on
+    the default direct solver of scipy sparse.
+
+    Parameters
+    ----------
+    nelx : int
+        number of elements in x direction.
+    nely : int
+        number of elements in y direction.
+    nelz : None or int
+        number of elements in z direction.
+    volfrac : float
+        volume fraction.
+    penal : float
+        penalty exponent for the SIMP method.
+    rmin : float
+        cutoff radius for the filter. Only elements within the element-center
+        to element center distance are used for filtering.
+    ft : int
+        integer flag for the filter. 0 sensitivity filtering,
+        1 density filtering, -1 no filter.
+    bcs : None or callable
+        boundary conditions.
+    lk : None or callable
+        element stiffness matrix.
+
+    Returns
+    -------
+    None.
+
+    """
+    #
+    if nelz is None:
+        ndim = 2
+        create_edofMat = create_edofMat2d
+    else:
+        ndim = 3
+        create_edofMat = create_edofMat3d
+    print("Minimum compliance problem with OC")
+    print("elements: "+" x ".join([f"{nelx}",f"{nely}",f"{nelz}"][:ndim]))
+    print("volfrac: " + str(volfrac) + ", rmin: " + str(rmin) + ", penal: " + str(penal))
+    print("Filter method: " + ["Sensitivity based","Density based"][ft])
+    # total number of design elements
+    n = np.prod([nelx,nely,nelz][:ndim])
+    # Max and min stiffness
+    Emin=1e-9
+    Emax=1.0
+    # Allocate design variables (as array), initialize and allocate sens.
+    x=volfrac * np.ones(n,dtype=float,order="F")
+    xold=x.copy()
+    xPhys=x.copy()
+    g=0 # must be initialized to use the NGuyen/Paulino OC approach
+    # fetch element stiffness matrix
+    KE = lk()
+    # dofs:
+    n_ndof = int(KE.shape[-1]/2**ndim)
[…]
