### 2025-10-01 d123d56
Added profiling.
