+++ examples/topology_optimization/compliance_minimization/heatplate2d.py
@@ -14,27 +14,34 @@ if __name__ == "__main__":
     ft = 0 # ft==0 -> sens, ft==1 -> dens
     display = True
     export = False
+    write_log = True
     #
     import sys
-    if len(sys.argv)>1: 
+    if len(sys.argv)>1:
         nelx = int(sys.argv[1])
-    if len(sys.argv)>2: 
+    if len(sys.argv)>2:
         nely = int(sys.argv[2])
-    if len(sys.argv)>3: 
+    if len(sys.argv)>3:
         volfrac = float(sys.argv[3])
-    if len(sys.argv)>4: 
+    if len(sys.argv)>4:
         rmin = float(sys.argv[4])
-    if len(sys.argv)>5: 
+    if len(sys.argv)>5:
         penal = float(sys.argv[5])
-    if len(sys.argv)>6: 
+    if len(sys.argv)>6:
         ft = int(sys.argv[6])
     if len(sys.argv)>7:
         display = bool(int(sys.argv[7]))
     if len(sys.argv)>8:
         export = bool(int(sys.argv[8]))
+    if len(sys.argv)>9:
+        write_log = bool(int(sys.argv[9]))
     #
     main(nelx=nelx, nely=nely, volfrac=volfrac, penal=penal, rmin=rmin, 
          ft=ft, filter_mode="matrix", optimizer="oc",nouteriter=1000,
          bcs=heatplate_2d, lk=lk_poisson_2d,
-         file = "heatplate2d",
-         debug=False, display=display, export=export)
+         output_kw = {"file": "heatplate_2d",
+                      "display": display,
+                      "export": export,
+                      "write_log": write_log,
+                      "profile": False,
+                      "debug": 0})
