+++ examples/topology_optimization/compliant_mechanisms/force_inverter_control2d.py
@@ -16,6 +16,7 @@ if __name__ == "__main__":
     ft = 0 # ft==0 -> sens, ft==1 -> dens
     display = True
     export = False
+    write_log = True
     #
     import sys
     if len(sys.argv)>1:
@@ -34,6 +35,8 @@ if __name__ == "__main__":
         display = bool(int(sys.argv[7]))
     if len(sys.argv)>8:
         export = bool(int(sys.argv[8]))
+    if len(sys.argv)>9:
+        write_log = bool(int(sys.argv[9]))
     #
     l = zeros((2*(nelx+1)*(nely+1),1))
     l[2*nelx*(nely+1),0] = 1
@@ -44,4 +47,9 @@ if __name__ == "__main__":
          bcs=forceinverter_2d, obj_func=var_squarederror,
          obj_kw={"l": l,"u0": u0},
          ft=ft, filter_mode="matrix",optimizer="ocm",
-         display=display,export=export)
+         output_kw = {"file": "force-inverter_control2d",
+                      "display": display,
+                      "export": export,
+                      "write_log": write_log,
+                      "profile": False,
+                      "debug": 0})
