+++ examples/topology_optimization/compliance_minimization/mbb_beam3d.py
@@ -14,6 +14,7 @@ if __name__ == "__main__":
     ft = 1 # ft==0 -> sens, ft==1 -> dens
     display = False
     export = False
+    write_log=True
     #
     import sys
     if len(sys.argv)>1: 
@@ -34,15 +35,21 @@ if __name__ == "__main__":
         display = bool(int(sys.argv[8]))
     if len(sys.argv)>9:
         export = bool(int(sys.argv[9]))
+    if len(sys.argv)>10:
+        write_log = bool(int(sys.argv[10])) 
     #
     main(nelx=nelx, nely=nely, nelz=nelz, volfrac=volfrac, penal=penal, 
          rmin=rmin, ft=ft, filter_mode="matrix", 
-         optimizer="oc", lin_solver = "scipy-direct",
+         optimizer="oc", lin_solver = "cvxopt-cholmod",
+         assembly_mode="lower",
          nouteriter=5,
          bcs=mbb_3d,
-         file="mbb_3d",
-         debug=False,display=display,
-         export=export)
+         output_kw = {"file": "mbb_3d",
+                      "display": display,
+                      "export": export,
+                      "write_log": write_log,
+                      "profile": True,
+                      "debug": 0})
     #main(nelx=nelx, nely=nely, nelz=nelz, volfrac=volfrac, penal=penal, rmin=rmin, 
     #     ft=ft, filter_mode="convolution", optimizer="oc",nouteriter=1000,
     #     bcs=mbb_3d,
