+++ examples/topology_optimization/compliance_minimization/cantilever_multiload2d.py
@@ -7,13 +7,15 @@ from topoptlab.objectives import compliance,compliance_squarederror
 if __name__ == "__main__":
     # Default input parameters
     nelx = 150
-    nely = 150
+    nely = nelx
     volfrac = 0.4
-    rmin = 6  # 5.4
+    rmin = 6.  # 5.4
     penal = 3.0
     ft = 0 # ft==0 -> sens, ft==1 -> dens
     display = True
     export = False
+    write_log = True
+    #
     import sys
     if len(sys.argv)>1:
         nelx = int(sys.argv[1])
@@ -31,6 +33,8 @@ if __name__ == "__main__":
         display = bool(int(sys.argv[7]))
     if len(sys.argv)>8:
         export = bool(int(sys.argv[8]))
+    if len(sys.argv)>9:
+        write_log = bool(int(sys.argv[9]))
     #
     main(nelx=nelx, nely=nely, volfrac=volfrac, penal=penal,
          rmin=rmin, ft=ft, filter_mode = "matrix",
@@ -38,4 +42,9 @@ if __name__ == "__main__":
          #obj_func=compliance_squarederror, obj_kw={"c0": [75, 80]},
          optimizer="mma",lin_solver="cvxopt-cholmod",
          bcs=cantilever_2d_twoloads_wrong,
-         debug=False,display=display,export=export)
+         output_kw = {"file": "cantilever_multiload2d",
+                      "display": display,
+                      "export": export,
+                      "write_log": write_log,
+                      "profile": False,
+                      "debug": 0})
