### 2025-10-02 99c21a7
Found bugs in AMG and GMG. Both fixed now. 3D needs still to be tested.
