+++ monolithic_codes/topopt88m.py
@@ -1,7 +1,7 @@
 # A 165 LINE TOPOLOGY OPTIMIZATION CODE BY NIELS AAGE AND VILLADS EGEDE JOHANSEN, JANUARY 2013
 # minor modifications by Stefan Hiemer (January 2025)
 import numpy as np
-from scipy.sparse import coo_matrix
+from scipy.sparse import coo_matrix,coo_array
 from scipy.sparse.linalg import spsolve,factorized
 from matplotlib import colors
 import matplotlib.pyplot as plt
@@ -104,7 +104,7 @@ def main(nelx,nely,volfrac,penal,rmin,ft,solver="lu"):
         loop=loop+1
         # Setup and solve FE problem
         sK=((KE.flatten()[np.newaxis]).T*(Emin+(xPhys)**penal*(Emax-Emin))).flatten(order='F')
-        K = coo_matrix((sK,(iK,jK)),shape=(ndof,ndof)).tocsc()
+        K = coo_array((sK,(iK,jK)),shape=(ndof,ndof)).tocsc()
         # add springs to stiffness matrix
         K[din,din] += kin # 
         K[dout,dout] += kout # 
