### 2025-10-03 33c04c6
Added PyAMG and got rid of some bugs in gmg. Some tests for pyamg still fail and need to write some for preconditioners only suited for Poisson-like problems.
