+++ topoptlab/gmg.py
@@ -175,7 +175,7 @@ def create_coarse_inds(nelx: int, nely: int, nelz: Union[None,int] = None,
         n = (ndof, nely+1, nelx+1)
         ndim = 2
     else:
-        n = (ndof, nelz+1, nelx+1, nely+1)
+        n = (ndof, nely+1, nelx+1, nelz+1)
         ndim = 3
     # convert stride to tuple
     if isinstance(stride,int):
@@ -186,7 +186,7 @@ def create_coarse_inds(nelx: int, nely: int, nelz: Union[None,int] = None,
     if nelz is None:
         return idx[:,::stride[1],::stride[0]].flatten(order="F")
     else:
-        return idx[:,::stride[2],::stride[0],::stride[1]].flatten(order="F")
+        return idx[:,::stride[1],::stride[0],::stride[2]].flatten(order="F")
 
 def create_coarse_mask(nelx: int, nely: int, nelz: Union[None,int] = None,
                        ndof: int = 1, stride: int = 2) -> np.ndarray:
