### 2025-10-03 3a1dabf
forgot to delete some debugging code.
