### 2025-10-03 f4eb20d
pyAMG is now included with tests. Need to include option to pass arguments for preconditioners to linsolve and through main().
