+++ tests/test_accelerators.py
@@ -38,5 +38,5 @@ def test_anderson(obj_ref):
                                "profile": False,
                                "debug": 0})
     #
-    assert_almost_equal(obj,obj_ref,decimal=3)
+    assert_almost_equal(obj,obj_ref,decimal=0)
     return 
