### 2025-10-05 5986a92
Rewrote the logging and added several details to the iterative solvers.
