+++ examples/topology_optimization/compliance_minimization/slab_uniform_strain_2d.py
@@ -50,4 +50,4 @@ if __name__ == "__main__":
                               "export": export,
                               "write_log": write_log,
                               "profile": False,
-                              "debug": 0})
+                              "verbosity": 20})
