+++ topoptlab/topology_optimization.py
@@ -37,7 +37,7 @@ from topoptlab.output_designs import export_vtk
 # map element data to img/voxel
 from topoptlab.utils import map_eltoimg,map_imgtoel,map_eltovoxel,map_voxeltoel,check_simulation_params
 # logging related stuff
-from topoptlab.log_utils import init_logging
+from topoptlab.log_utils import EmptyLogger,SimpleLogger
 #
 from mmapy import mmasub
 
@@ -73,7 +73,7 @@ def main(nelx: int, nely: int,
                             "export": True,
                             "write_log": True,
                             "profile": False,
-                            "debug": 0}) -> Tuple[np.ndarray,float]:
+                            "verbosity": 20}) -> Tuple[np.ndarray,float]:
     """
     Topology optimization workflow with the material interpolation method. 
     Can treat single physics stationary problems.
@@ -164,23 +164,28 @@ def main(nelx: int, nely: int,
     #
     if output_kw["write_log"]:
         # check if log file exists and if True delete
-        to_log = init_logging(logfile=output_kw["file"])
+        log = SimpleLogger(file=output_kw["file"],
+                           verbosity=output_kw["verbosity"])
+        
         #
-        to_log(f"optimizer {optimizer}")
-        to_log(f"number of spatial dimensions: {ndim}")
-        to_log("elements: "+" x ".join([f"{nelx}",f"{nely}",f"{nelz}"][:ndim]))
+        log.info(f"optimizer {optimizer}")
+        log.info(f"number of spatial dimensions: {ndim}")
+        log.info("elements: "+" x ".join([f"{nelx}",f"{nely}",f"{nelz}"][:ndim]))
         if volfrac is not None:
-            to_log(f"volfrac: {volfrac} rmin: {rmin}  penal: {penal}")
+            log.info(f"volfrac: {volfrac} rmin: {rmin}  penal: {penal}")
         else:
-            to_log(f"rmin: {rmin}  penal: {penal}")
-        to_log("filter: " + ["Sensitivity based",
+            log.info(f"rmin: {rmin}  penal: {penal}")
+        log.info("filter: " + ["Sensitivity based",
                              "Density based",
                              "Haeviside Guest",
                              "Haeviside complement Sigmund 2007",
                              "Haeviside eta projection",
                              "Volume Preserving eta projection",
                              "No filter"][ft])
-        to_log(f"filter mode: {filter_mode}")
+        log.info(f"filter mode: {filter_mode}")
+    else:
+        # check if log file exists and if True delete
+        log = EmptyLogger()
     # total number of design elements
     n = np.prod([nelx,nely,nelz][:ndim])
     #
@@ -468,8 +473,8 @@ def main(nelx: int, nely: int,
                 if "density_coupled" in body_forces_kw.keys():
                     dobj[:,0] -= simp_dx(xPhys=xPhys, eps=0., penal=1.)[:,0]*\
                                          np.dot(adj[edofMat,i],fe_dens[:,i])
-                if output_kw["debug"]:
-                    print("FEM: it.: {0}, problem: {1}, min. u: {2:.10f}, med. u: {3:.10f}, max. u: {4:.10f}".format(
+                #
+                log.debug("[DEBUG] FEM: it.: {0}, problem: {1}, min. u: {2:.10f}, med. u: {3:.10f}, max. u: {4:.10f}".format(
                            loop,i,np.min(u[:,i]),np.median(u[:,i]),np.max(u[:,i])))
[…]
