+++ examples/topology_optimization/compliance_minimization/heatplate2d.py
@@ -44,4 +44,4 @@ if __name__ == "__main__":
                       "export": export,
                       "write_log": write_log,
                       "profile": False,
-                      "debug": 0})
+                      "verbosity": 20})
