+++ examples/topology_optimization/compliance_minimization/mbb_beam3d.py
@@ -49,7 +49,7 @@ if __name__ == "__main__":
                       "export": export,
                       "write_log": write_log,
                       "profile": True,
-                      "debug": 0})
+                      "verbosity": 20})
     #main(nelx=nelx, nely=nely, nelz=nelz, volfrac=volfrac, penal=penal, rmin=rmin, 
     #     ft=ft, filter_mode="convolution", optimizer="oc",nouteriter=1000,
     #     bcs=mbb_3d,
