+++ examples/topology_optimization/compliance_minimization/mbb_beam2d.py
@@ -55,4 +55,4 @@ if __name__ == "__main__":
                               "export": export,
                               "write_log": write_log,
                               "profile": False,
-                              "debug": 0})
+                              "verbosity": 20})

[…]
