### 2025-10-06 7beaad1
Better logfile parsing.
