### 2025-10-07 fe8b867
Added rigid body modes for AMG for lin. elasticity problems. Also got some of the util functions more consistent.
