+++ examples/topology_optimization/compliance_minimization/slab_uniform_strain_2d.py
@@ -42,7 +42,7 @@ if __name__ == "__main__":
     #
     x,obj = main(nelx=nelx, nely=nely, volfrac=volfrac, penal=penal,
                  rmin=rmin, ft=ft, filter_mode="matrix",
-                 optimizer="mma", lin_solver="scipy-direct",
+                 optimizer="mma", lin_solver_kw={"name": "scipy-direct"},
                  nouteriter=1000,
                  bcs=singlenode, body_forces_kw={"strain_uniform": np.eye(3)},
                  output_kw = {"file": "uniformstrain_2d",
