+++ tests/test_utils.py
@@ -1,4 +1,4 @@
-from numpy import arange,array,prod
+from numpy import arange,array,column_stack,prod
 from numpy.testing import assert_almost_equal
 
 import pytest
@@ -45,10 +45,10 @@ from topoptlab.utils import elid_to_coords
 @pytest.mark.parametrize('nelx, nely, nelz, solution',
                          [(2,3,None,
                            (array([0, 0, 0, 1, 1, 1]), 
-                            array([0, 1, 2, 0, 1, 2]))),
+                            array([2, 1, 0, 2, 1, 0]))),
                           (2,3,2,
                            (array([0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1]), 
-                            array([0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2]), 
+                            array([2, 1, 0, 2, 1, 0, 2, 1, 0, 2, 1, 0]), 
                             array([0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1])))])
 
 def test_elidtocoords(nelx, nely, nelz, solution):
@@ -60,8 +60,63 @@ def test_elidtocoords(nelx, nely, nelz, solution):
     #
     elid = arange( prod( [nelx,nely,nelz][:ndim]).astype(int) )
     #
-    assert_almost_equal(solution,
-                        elid_to_coords(el=elid,nelx=nelx,nely=nely,nelz=nelz)) 
+    assert_almost_equal(elid_to_coords(el=elid,
+                                       nelx=nelx,nely=nely,nelz=nelz),
+                        column_stack(solution)) 
+    
+    return
+
+from topoptlab.utils import nodeid_to_coords
+
+@pytest.mark.parametrize('nelx, nely, nelz, solution',
+                         [(2,3,None,
+                           (array([-0.5, -0.5, -0.5,-0.5,
+                                    0.5, 0.5, 0.5, 0.5, 
+                                    1.5, 1.5, 1.5, 1.5]), 
+                            array([2.5, 1.5, 0.5, -0.5,
+                                   2.5, 1.5, 0.5, -0.5,
+                                   2.5,1.5,0.5,-0.5]))),
+                          (2,3,2,
+                           (array([-0.5, -0.5, -0.5, -0.5, 
+                                   0.5, 0.5, 0.5, 0.5,
+                                   1.5, 1.5, 1.5, 1.5,
+                                   -0.5, -0.5, -0.5, -0.5, 
+                                   0.5, 0.5, 0.5, 0.5,
+                                   1.5, 1.5, 1.5, 1.5,
+                                   -0.5, -0.5, -0.5, -0.5, 
+                                   0.5, 0.5, 0.5, 0.5,
+                                   1.5, 1.5, 1.5, 1.5]), 
+                            array([2.5, 1.5, 0.5, -0.5,
+                                   2.5, 1.5, 0.5, -0.5,
+                                   2.5, 1.5, 0.5, -0.5,
+                                   2.5, 1.5, 0.5, -0.5,
+                                   2.5, 1.5, 0.5, -0.5,
+                                   2.5, 1.5, 0.5, -0.5,
+                                   2.5, 1.5, 0.5, -0.5,
+                                   2.5, 1.5, 0.5, -0.5,
+                                   2.5, 1.5, 0.5, -0.5]), 
+                            array([-0.5, -0.5, -0.5, -0.5, 
+                                   -0.5, -0.5, -0.5, -0.5,
+                                   -0.5, -0.5, -0.5, -0.5, 
+                                   0.5, 0.5, 0.5, 0.5, 
+                                   0.5, 0.5, 0.5, 0.5, 
[…]
