### 2025-10-08 619c766
Added a cg without preconditioning for benchmarking, fixed some typos in the docu and included proper initial guess into topology_optimization.main().
