### 2025-10-10 72ec824
removed the ichol. it is too slow in that form and also is not useful as it does not improve convergence in many cases.
