+++ tests/test_solver.py
@@ -1,4 +1,5 @@
 from numpy.testing import assert_almost_equal
+from numpy import array
 
 import pytest
 
