### 2025-10-13 930c43f
added conversion between different stress measures for the symbolic part.
