### 2025-10-13 1d69902
added conversion between different stress measures for the symbolic part.
