+++ topoptlab/symbolic/strain_measures.py
@@ -193,24 +193,3 @@ def green_strain(ndim: int,
                  element_type=element_type,
                  order=order)
     return simplify_matrix(F@F.transpose())
-
-def is_equal(expr1: Expr, expr2: Expr)-> bool:
-    """
-    Return True if two SymPy expressions are algebraically identical.
-
-    The check expands the difference and asks SymPy to simplify it to zero.
-
-    Parameters
-    ----------
-    expr1 : sympy.Expr
-        first expression.
-    expr2 : sympy.Expr
-        second expression.
-
-    Returns
-    -------
-    bool
-        True if ``expr1 - expr2`` simplifies to zero (identical),
-        False otherwise.
-    """
-    return simplify(expand(expr1 - expr2)) == 0
