+++ topoptlab/symbolic/stress_conversions.py
@@ -1,0 +1,187 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+from typing import Union
+
+from sympy import expand,simplify,Expr
+from symfem.functions import ScalarFunction,MatrixFunction
+
+from topoptlab.symbolic.cell import base_cell
+from topoptlab.symbolic.matrix_utils import simplify_matrix, generate_constMatrix
+from topoptlab.symbolic.matrix_utils import inverse
+
+def cauchy_to_pk1(sigma: MatrixFunction, 
+                  F: MatrixFunction,
+                  Fdet: Union[None,ScalarFunction] = None) -> MatrixFunction:
+    """
+    Convert Cauchy stress `sigma` to first Piola-Kirchhoff stress `P`:
+        
+        `P = det(F) sigma@F^{-1}.T`
+
+    Parameters
+    ----------
+    sigma : symfem.functions.MatrixFunction
+        Cauchy stress tensor of shape (ndim,ndim).
+    F : symfem.functions.MatrixFunction
+        deformation gradient of shape (ndim,ndim).
+    Fdet : None or symfem.functions.ScalarFunction
+        determinant of deformation gradient.
+
+    Returns
+    -------
+    P : symfem.functions.MatrixFunction
+        first Piola-Kirchhoff stress tensor of shape (ndim,ndim).
+
+    """
+    #
+    if Fdet is None:
+        Fdet = F.det() 
+    #
+    return simplify_matrix( (sigma@inverse(F).transpose()).__mul__(Fdet) )
+
+def cauchy_to_pk2(sigma: MatrixFunction, 
+                  F: MatrixFunction,
+                  Fdet: Union[None,ScalarFunction] = None,
+                  Finv: Union[None,MatrixFunction] = None) -> MatrixFunction:
+    """
+    Convert Cauchy stress `sigma` to second Piola-Kirchhoff stress `S`:
+        
+        `S = det(F) F^{-1}@sigma@F^{-1}.T`
+
+    Parameters
+    ----------
+    sigma : symfem.functions.MatrixFunction
+        Cauchy stress tensor of shape (ndim,ndim).
+    F : symfem.functions.MatrixFunction
+        deformation gradient of shape (ndim,ndim).
+    Fdet : None or symfem.functions.ScalarFunction
+        determinant of deformation gradient.
+    Finv : None or symfem.functions.MatrixFunction
+        inverse of deformation gradient.
+
+    Returns
+    -------
+    S : symfem.functions.MatrixFunction
+        second Piola-Kirchhoff stress tensor of shape (ndim,ndim).
+
+    """
+    #
+    if Fdet is None:
+        Fdet = F.det()
+    #
+    if Finv is None:
+        Finv = inverse(F)
+    #
+    return simplify_matrix((Finv@sigma@Finv.transpose()).__mul__(Fdet))
+
+def pk1_to_cauchy(P: MatrixFunction, 
+                  F: MatrixFunction,
+                  Fdet: Union[None,ScalarFunction] = None) -> MatrixFunction:
+    """
+    Convert first Piola-Kirchhoff stress `P` to Cauchy stress `sigma`:
+        
+        `sigma = P@F.T / det(F)`
+
+    Parameters
+    ----------
+    P : symfem.functions.MatrixFunction
+        first Piola-Kirchhoff stress tensor of shape (ndim,ndim).
+    F : symfem.functions.MatrixFunction
+        deformation gradient of shape (ndim,ndim).
+    Fdet : None or symfem.functions.ScalarFunction
+        determinant of deformation gradient.
+
+    Returns
+    -------
+    sigma : symfem.functions.MatrixFunction
[…]
