+++ topoptlab/filter/haeviside_projectors.py
@@ -1,0 +1,347 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+from typing import Any, Union
+
+import numpy as np
+
+from topoptlab.filter.filter import TOFilter
+from topoptlab.filter.haeviside_projection import find_eta
+
+class HaevisideProjectorGuest2004(TOFilter):
+    """
+    
+    Implements the Haeviside projection by 
+    
+    Guest, James K., Jean H. Prévost, and Ted Belytschko. "Achieving minimum 
+    length scale in topology optimization using nodal design variables and 
+    projection functions." International journal for numerical methods in 
+    engineering 61.2 (2004): 238-254.
+    
+    This projection is a smooth version of the Haeviside step function Theta(x), 
+    so in simple words, this projection sets every value that is larger than 
+    zero to one and everything smaller/equal to zero to zero. The filter 
+    equation is 
+    
+    x_filtered = 1 - exp(-beta x) + x exp(-beta)
+    
+    beta is the projection strength, that is typically ramped up during the TO 
+    process to large values. The larger beta, the closer this filter is to a 
+    Haeviside function. 
+    """
+    
+    def __init__(self,
+                 **kwargs: Any) -> None:
+        """
+        Do nothing as this filter requires no initialization.
+        
+        Parameters
+        ----------
+        None
+        
+        Returns
+        -------
+        None
+
+        """
+        return
+        
+    def apply_filter(self, 
+                     x: np.ndarray, 
+                     beta=float,
+                     **kwargs: Any) -> np.ndarray:
+        """
+        Apply filter to (intermediate) design variables x
+        
+        x_filtered = 1 - exp(-beta x) + x exp(-beta)
+        
+        Parameters
+        ----------
+        x : np.ndarray
+            (intermediate) design variables.
+        beta : float
+            projection strength.
+
+        Returns
+        -------
+        x_filtered : np.ndarray
+            filtered design variables.
+
+        """
+        return 1 - np.exp(-beta*x) + x*np.exp(-beta)
+    
+    def apply_filter_dx(self, 
+                        x_filtered : np.ndarray, 
+                        beta : float,
+                        **kwargs: Any) -> np.ndarray:
+        """
+        Apply filter to the sensitivities with respect to filtered variables 
+        x_filtered using the chain rule assuming 
+        
+            x_filtered = 1 - exp(-beta x) + x exp(-beta) 
+        
+        to get the sensitivities with respect to the (unfiltered) design 
+        variables or in the case of many filters intermediate design variables:
+            
+            dx = beta*np.exp(-beta*x_filtered) + np.exp(-beta)
+        
+        Parameters
+        ----------
+        x_filtered : np.ndarray
+            filtered design variables.
+        beta : float
+            projection strength.
+            
+        Returns
+        -------
+        dx : np.ndarray
+            design sensitivities with respect to un-filtered design variables.
+        """
+        return beta*np.exp(-beta*x_filtered) + np.exp(-beta)
+    
+    @property
+    def vol_conserv(self) -> bool:
+        """
+        Set self.vol_conserv to False as filter is not volume conserving. 
+        
+        Parameters
+        ----------
+        None.
+            
+        Returns
+        -------
+        False
+        """
+        return False
+
+class HaevisideProjectorSigmund2007(TOFilter):
+    """
+    
+    Implements the Haeviside projection by 
+    
+    Sigmund, Ole. "Morphology-based black and white filters for topology 
+    optimization." Structural and Multidisciplinary Optimization 33.4 (2007): 
+    401-424.
+    
+    This projection is a different flavor of the Guest2004 projection and 
+    approximates the shifted Haeviside step function Theta(1-x), so in this 
[…]
