### 2025-10-14 9d2900b
Wrapped filters in classes. Now needs to be included in the main topology optimization routine. Also ideally I should separate the individual filter functions out to make the Matlab-like user pattern possible.
