+++ topoptlab/solve_linsystem.py
@@ -97,11 +97,6 @@ def solve_lin(K: Union[csc_array,spmatrix], rhs: Union[np.ndarray,matrix],
             ilu = spilu(K, **preconditioner_kw)
             P = LinearOperator(shape=K.shape,
                                matvec=ilu.solve)
-        elif preconditioner == "topoptlab-ichol":
-            Lhat=ichol(A=K)
-            P = LinearOperator(shape=K.shape,
-                               matvec=partial(apply_ichol,
-                                              Lhat=Lhat))
         elif preconditioner == "pyamg-air":
             P = air_solver(A=K,
                            **preconditioner_kw).aspreconditioner(cycle='V')
