+++ examples/topology_optimization/compliance_minimization/heatplate2d.py
@@ -7,13 +7,13 @@ from topoptlab.elements.poisson_2d import lk_poisson_2d
 if __name__ == "__main__":
     # Default input parameters
     nelx = 40
-    nely = 40
+    nely = nelx
     volfrac = 0.4
-    rmin = 1.2
+    rmin = 0.03*nelx
     penal = 3.0
     ft = 0 # ft==0 -> sens, ft==1 -> dens
-    display = True
-    export = False
+    display = False
+    export = True
     write_log = True
     #
     import sys
@@ -38,6 +38,8 @@ if __name__ == "__main__":
     #
     main(nelx=nelx, nely=nely, volfrac=volfrac, penal=penal, rmin=rmin, 
          ft=ft, filter_mode="matrix", optimizer="oc",nouteriter=1000,
+         #lin_solver_kw = {"name": "cvxopt-cholmod"},
+         #lin_solver_kw = {"name": "topoptlab-cg"}, preconditioner_kw = {"name": "pyamg-pyamg-ruge_stuben"},
          bcs=heatplate_2d, lk=lk_poisson_2d,
          output_kw = {"file": "heatplate_2d",
                       "display": display,
