### 2025-10-23 e947a9d
fixed a number of typos and an error in the final thresholding that stemmed from a shape inconsistency in the physical densities.
