### 2025-10-24 c11f63a
Minor edits in docu., some bugfixes and extension of the heaplate case to 3d.
