### 2025-10-24 af49287
Minor edits in docu., some bugfixes and extension of the heaplate case to 3d.
