### 2025-11-01 591a6f6
kicked out symmetry bc for heat plate as it does not make sense there. fixed 2 bugs in tests.
