+++ examples/topology_optimization/compliance_minimization/heatplate2d.py
@@ -1,6 +1,4 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
-from functools import partial
-
 from topoptlab.topology_optimization import main
 from topoptlab.example_bc.heat_conduction import heatplate_2d
 from topoptlab.elements.poisson_2d import lk_poisson_2d
@@ -42,7 +40,7 @@ if __name__ == "__main__":
          ft=ft, filter_mode="matrix", optimizer="oc",nouteriter=1000,
          #lin_solver_kw = {"name": "cvxopt-cholmod"},
          #lin_solver_kw = {"name": "topoptlab-cg"}, preconditioner_kw = {"name": "pyamg-pyamg-ruge_stuben"},
-         bcs=partial(heatplate_2d,symmetry=1), 
+         bcs=heatplate_2d, 
          lk=lk_poisson_2d,
          output_kw = {"file": "heatplate_2d",
                       "display": display,
