+++ topoptlab/example_bc/heat_conduction.py
@@ -3,8 +3,7 @@ from typing import Any,Tuple
 import numpy as np
 
 def heatplate_2d(nelx: int, nely: int,
-                 ndof: int, 
-                 symmetry: bool = True,
+                 ndof: int,
                  **kwargs: Any
                  ) -> Tuple[np.ndarray,np.ndarray,np.ndarray,np.ndarray,None]:
     """
@@ -20,8 +19,6 @@ def heatplate_2d(nelx: int, nely: int,
         number of elements in y direction.
     ndof : int
         number of degrees of freedom.
-    symmetry : bool
-        if True, creates a mirror symmetry condition.
 
     Returns
     -------
@@ -44,17 +41,15 @@ def heatplate_2d(nelx: int, nely: int,
     u = np.zeros((ndof, 1))
     # load/source
     f[:, 0] = 0.01 # constant source
-    if not symmetry:
-        fixed = np.arange(int(nely / 2 - nely / 20), 
-                          int(nely / 2 + 1 + nely / 20))
-    else:
-        fixed = np.arange( 1+int(nely / 10))
-        f[np.arange(0,(nelx+1)*(nely+1),(nely+1)),0] = 0.
+    fixed = np.arange(int(nely / 2 - nely / 20), 
+                      int(nely / 2 + 1 + nely / 20))
+    # symmetry condition, but nonsense for this example
+    #fixed = np.arange( 1+int(nely / 10))
+    #f[np.arange(0,(nelx+1)*(nely+1),(nely+1)),0] = 0.
     return u,f,fixed,np.setdiff1d(dofs,fixed),None
 
 def heatplate_3d(nelx: int, nely: int, nelz: int,
                  ndof: int, 
-                 symmetry: bool = True,
                  **kwargs: Any
                  ) -> Tuple[np.ndarray,np.ndarray,np.ndarray,np.ndarray,None]:
     """
@@ -72,8 +67,6 @@ def heatplate_3d(nelx: int, nely: int, nelz: int,
         number of elements in z direction.
     ndof : int
         number of degrees of freedom.
-    symmetry : bool
-        if True, creates two mirror symmetry conditions.
 
     Returns
     -------
@@ -96,16 +89,15 @@ def heatplate_3d(nelx: int, nely: int, nelz: int,
     u = np.zeros((ndof, 1))
     # load/source
     f[:, 0] = 0.01 # constant source
-    if not symmetry:
-        fixed = np.arange(int(nely / 2 + 1 - nely / 20)-1, 
-                          int(nely / 2 + 1 + nely / 20))
-    else:
-        fixed = np.arange( 1+int(nely / 10))
-        #
-        ysymmetry = np.arange(0,ndof,(nely+1))
-        # symmetry bc (fix z displacements to zero)
-        zsymmetry = np.arange((nelx+1)*(nely+1))
-        f[np.hstack( (ysymmetry,zsymmetry)),0] = 0.
+    fixed = np.arange(int(nely / 2 + 1 - nely / 20)-1, 
+                      int(nely / 2 + 1 + nely / 20))
+    # symmetry but nonsense for this rudimentary bcs
+    fixed = np.arange( 1+int(nely / 10))
+    # symmetry bc (fix z displacements to zero)
+    #ysymmetry = np.arange(0,ndof,(nely+1))
+    # 
+    #zsymmetry = np.arange((nelx+1)*(nely+1))
+    #f[np.hstack( (ysymmetry,zsymmetry)),0] = 0.
     #
     fixed = np.tile(fixed, nelz+1)+np.repeat(np.arange(nelz+1), 
                                              fixed.shape)*(nelx+1)*(nely+1)
