### 2025-11-01 d3dabbe
kicked out symmetry bc for heat plate as it does not make sense there. fixed 2 bugs in tests.
