### 2025-11-02 c809a06
added symbolic lin. elasticity.
