### 2025-11-03 0cc84a9
incomplete crash course to FEM added to documentation. need to extract nonlin. mechanics to another tutorial.
