+++ docs/source/topoptlab.blocksparse_precond.rst
@@ -1,7 +1,0 @@
-topoptlab.blocksparse\_precond module
-=====================================
-
-.. automodule:: topoptlab.blocksparse_precond
-   :members:
-   :show-inheritance:
-   :undoc-members:
