+++ docs/source/introduction/crashcourse_fem.md
@@ -1,0 +1,144 @@
+# A Crashcourse to FEM
+Given the limitations in space and time, this is a brief introduction to the 
+Finite Element Method (FEM). Interested readers are referred to the classic 
+textbook by Belytschko and Fish {cite}`fish2007first`, the excellent lecture 
+notes by Dennis Kochmann (https://mm.ethz.ch/education/lecture-notes.html), and 
+the textbook by Peter Wriggers {cite}`wriggers2008nonlinear` for advanced 
+topics. This section outlines the basic "recipe" to discretize a physical 
+problem described by a partial differential equation (PDE) using FEM.
+
+## From Strong to Weak Form
+Most physical problems can be written as PDEs based on a conservation law or 
+other physical insights. Classical examples are the **Poisson equation**
+
+```{math}
+-\nabla \cdot (\boldsymbol{K} \nabla \phi) = f
+```
+governing heat conduction or diffusion with scalar variable $\phi$ 
+(temperature, concentration, etc.) and property tensor $\boldsymbol{K}$, 
+the continuity equation
+```{math}
+\frac{\partial \rho}{\partial t} + \nabla \cdot (\rho \boldsymbol{v}) = 0
+```
+describing conservation of mass (density $\rho$, time $t$ and velocity field 
+$\boldsymbol{v}$ and the **Cauchy momentum balance**
+```{math}
+\nabla \cdot \boldsymbol{\sigma} + \boldsymbol{b} = \rho \ddot{\boldsymbol{u}}
+```
+representing conservation of linear momentum with the Cauchy stress tensor 
+$\boldsymbol{\sigma}$, body force $\boldsymbol{b}$ and displacement 
+$\boldsymbol{u}$ . Combined with a constitutive law (e.g. Hooke’s law 
+$\boldsymbol{\sigma} = \boldsymbol{C} : \boldsymbol{\epsilon}$ with stiffness 
+tensor $\boldsymbol{C}$ and engineering/incremental strain tensor
+$\boldsymbol{\epsilon}=\frac{1}{2}\left(\nabla \boldsymbol{u} + \nabla \boldsymbol{u}^T \right)$), this gives the mechanical response of a solid. For 
+small deformations, inserting Hooke’s law yields the **Navier–Lamé equations**
+```{math}
+\nabla \cdot \left( \boldsymbol{C} : \boldsymbol{\epsilon} \right)  + \boldsymbol{b} = \rho \ddot{\boldsymbol{u}},
+```
+and for Newtonian fluids the **Navier–Stokes equations**
+```{math}
+\rho \left( \frac{\partial \boldsymbol{v}}{\partial t} + \boldsymbol{v}\cdot\nabla\boldsymbol{v} \right)
+= -\nabla p + \mu \nabla^2 \boldsymbol{v} + \rho \boldsymbol{b}.
+```
+
+All of these equations are expressed in terms of their derivatives i. e. 
+**differential form** which is also called the **strong form**. We introduce a 
+generic PDE for brevity as
+```{math}
+\mathcal{L}(u) = f \quad \text{in } \Omega,
+```
+where the linear differential operator $\mathcal{L}$ represents all terms with 
+derivatives involving the unknown field $u$ (e. g. for linear elasticity 
+$\nabla \cdot \left( \boldsymbol{C} : \boldsymbol{\epsilon} \right)$), 
+$f$ represents any inhomogeneous terms often called source/sink terms and 
[…]
