### 2025-11-04 0d2c0da
small edits/fixes in docu.
