### 2025-11-05 d9bef1a
small edits/fixes in docu.
