+++ docs/source/introduction/crashcourse_fem.md
@@ -18,9 +18,8 @@ only hint or show final results for nonlinear problems.
 ## From Strong to Weak Form
 Most physical problems can be written as PDEs based on a conservation law or 
 other physical insights. Classical examples are the **Poisson equation**
-
 ```{math}
--\nabla \cdot (\boldsymbol{K} \nabla \phi) = f
+\nabla \cdot (\boldsymbol{K} \nabla \phi) = f
 ```
 governing heat conduction or diffusion with scalar variable $\phi$ 
 (temperature, concentration, etc.) and property tensor $\boldsymbol{K}$, 
@@ -31,24 +30,24 @@ the continuity equation
 describing conservation of mass (density $\rho$, time $t$ and velocity field 
 $\boldsymbol{v}$ and the **Cauchy momentum balance**
 ```{math}
-\nabla \cdot \boldsymbol{\sigma} + \boldsymbol{b} = \rho \ddot{\boldsymbol{u}}
+\rho \ddot{\boldsymbol{u}} = \nabla \cdot \boldsymbol{\sigma} + \boldsymbol{b}
 ```
 representing conservation of linear momentum with the Cauchy stress tensor 
 $\boldsymbol{\sigma}$, body force $\boldsymbol{b}$ and displacement 
 $\boldsymbol{u}$ . Combined with a constitutive law (e.g. Hooke’s law 
 $\boldsymbol{\sigma} = \boldsymbol{C} : \boldsymbol{\epsilon}$ with stiffness 
 tensor $\boldsymbol{C}$ and engineering/incremental strain tensor
-$\boldsymbol{\epsilon}=\frac{1}{2}\left(\nabla \boldsymbol{u} + \nabla \boldsymbol{u}^T \right)$), this gives the mechanical response of a solid. For 
-small deformations, inserting Hooke’s law yields the **Navier–Lamé equations**
+$\boldsymbol{\epsilon}=\frac{1}{2}\left(\nabla \boldsymbol{u} + \nabla \boldsymbol{u}^T \right)$), 
+this gives the mechanical response of a solid. For small deformations, 
+inserting Hooke’s law yields the **Navier–Lamé equations**
 ```{math}
-\nabla \cdot \left( \boldsymbol{C} : \boldsymbol{\epsilon} \right)  + \boldsymbol{b} = \rho \ddot{\boldsymbol{u}},
+\rho \ddot{\boldsymbol{u}} = \nabla \cdot \left( \boldsymbol{C} : \boldsymbol{\epsilon} \right)  + \boldsymbol{b},
 ```
 and for Newtonian fluids the **Navier–Stokes equations**
 ```{math}
 \rho \left( \frac{\partial \boldsymbol{v}}{\partial t} + \boldsymbol{v}\cdot\nabla\boldsymbol{v} \right)
 = -\nabla p + \mu \nabla^2 \boldsymbol{v} + \rho \boldsymbol{b}.
 ```
-
 All of these equations are expressed in terms of their derivatives i. e. 
 **differential form** which is also called the **strong form**. We introduce a 
 generic PDE for brevity as
@@ -59,26 +58,64 @@ where the linear differential operator $\mathcal{L}$ represents all terms with
 derivatives involving the unknown field $u$ (e. g. for linear elasticity 
 $\nabla \cdot \left( \boldsymbol{C} : \boldsymbol{\epsilon} \right)$), 
 $f$ represents any inhomogeneous terms often called source/sink terms and 
[…]
