### 2025-11-07 f48bfd2
small changes to docu.
