+++ docs/source/introduction/crashcourse_fem.md
@@ -79,20 +79,24 @@ Furthermore, from a numerical perspective, it is often convenient to work with
 integrals as they can be approximated cheaply with low-order splines or 
 polynomials.
 
-### Example 1: Poisson equation / Heat conduction / Diffusion
-We write down weak form via standard procedure 
-i) multiply by $w$ ii) integrate over domain $\Omega$:
+### Example: Weak form of  Poisson equation 
+In this paragraph we will write down the weak form for the Poisson equation
+which guides an abundant number of physical phenomena like temperature 
+conduction, diffusion, gravity, electrostatics, etc. pp. For sake of brevity, 
+we focus on time-independent (stationary) heat conduction. We start with the 
+weak form via standard procedure i) multiply by $w$ ii) integrate over domain 
+$\Omega$:
 ```{math}
 \int_\Omega w \nabla \cdot (\boldsymbol{K} \nabla \phi) dV = \int_\Omega w f dV
 ```
-Technically one can stop now as this is correct, but we would like to reduce 
-the highest order derivative as much as we can as in FEM this means we can 
-approximate it cheaper. We split weak form into left hand side and right hand 
-side. Now consider the "problematic" left hand side
+Technically one can stop now as this is a correct weak from, but we want to 
+reduce the highest order derivative as much as we can as in FEM this means we 
+can approximate it cheaper. We split weak form into left hand side and right 
+hand side and consider the "problematic" left hand side
 ```{math}
 \int_\Omega w \nabla \cdot (\boldsymbol{K} \nabla \phi) dV
 ```
-and try to simplify it. We write down the chain rule for a general vector $\boldsymbol{v}$ 
+to simplify it further. We remember the chain rule for a general vector $\boldsymbol{v}$ 
 and the scalar function $w$
 ```{math}
 \nabla \cdot (w\boldsymbol{v}) = w\nabla \cdot \boldsymbol{v} + \nabla w \cdot \boldsymbol{v}
@@ -106,90 +110,59 @@ We then insert $\boldsymbol{v}=\nabla \phi$ and rewrite the left hand side to
 \int_\Omega w \nabla \cdot (\boldsymbol{K} \nabla \phi) dV = \int_\Omega \nabla w \cdot \boldsymbol{K} \nabla \phi dV - \int_\Omega \nabla \cdot \left(w \boldsymbol{K} \nabla \phi\right) dV
 ```
 If we inspect this closer, we recognize that the second term on the right hand 
-side is the volume integral of the divergence of the flow $\boldsymbol{K} \nabla \phi$ 
-scaled by $w$. 
-divergence theorem
+side is the volume integral of the divergence of the flow 
+$\boldsymbol{K} \nabla \phi$ scaled by $w$. In simple words, this integral 
+describes how much of $w \boldsymbol{K} \nabla \phi $ is being produced or lost 
+within the volume. Instead of measuring what happens inside, we can 
+equivalently measure how much of $w \boldsymbol{K} \nabla \phi $ flows in or 
+out through the surface enclosing it which is what the divergence theorem 
+expresses
 ```{math}
[…]
