### 2025-11-12 stefanhiemer/topoptlab#3 @tisaac
More details on comparing to the state of the field in paper.md
In openjournals/joss-reviews#9105, I'm asked to review your comparison of topoptlab to other options in the field.  Your prose description in "Statement of need" section is fine, but it feels incomplete without some code examples.

Can you please add to your paper:

- An example of a call to `topology_optimization` problem to solve equations like those in the "Summary" section.
- An example of how one goes about getting "access to low-level implementation details which is necessary for research in TO," since this is supposed to be a comparative strength of topoptlab w.r.t. other packages.
