### 2025-11-12 5d644bc
small changes to docu and started to implement symbolic nonlinear heat conduction.
