### 2025-11-14 5efe71b
Periodic boundary conditions for matrix filter.
