### 2025-11-17 c96406a
Documentation of all examples via the README and clean-up.
