+++ examples/linsolvers/laplacian.py
@@ -61,7 +61,7 @@ def amg_solver(A, b,
                         A = A, b = b, 
                         interpolators= interpolators,
                         cycle = vcycle, 
-                        tol=tol,
+                        tol = tol,
                         smoother_fnc=smoothed_jacobi,
                         smoother_kws=smoother_kws,
                         max_cycles=max_cycles,
