+++ examples/linsolvers/show_coarse.py
@@ -8,7 +8,7 @@ from scipy.sparse import load_npz
 
 from topoptlab.amg import standard_coarsening
 from topoptlab.gmg import create_coarse_mask
-from topoptlab.blocksparse_precond import create_primitive_blocks
+from topoptlab.blocksparse.make_blocks import create_equal_blocks
 from topoptlab.utils import map_eltoimg
 from topoptlab.example_bc.lin_elast import mbb_2d
 from topoptlab.example_bc.heat_conduction import heatplate_2d
@@ -58,7 +58,7 @@ if __name__ == "__main__":
     # algebraic multigrid
     amg_mask = standard_coarsening(A=L)
     # block_precond
-    block_masks = create_primitive_blocks(A=L,nblocks=2)
+    block_masks = create_equal_blocks(A=L,nblocks=2)
     #
     fig,axs = plt.subplots(1,3)
     #
