+++ examples/fem/README.md
@@ -1,0 +1,15 @@
+These directories contain different FEM demonstration cases. 
+
+linear_elasticity
+
+Standard stationary, small-strain, small-rotations linear elasticity cases.
+
+cahn-hilliard
+
+Solves transient Cahn-Hilliard equation via semi-implicite, mixed FEM.
+
+heat_expansion
+
+Stationary, small-strain, small-rotations linear elasticity coupled with heat expansion
+due to a uniform or non-uniform change in temperature.
+

[…]
