+++ examples/README.md
@@ -1,36 +1,48 @@
 These directories contain application cases of the topoptlab. 
 
-elements
+## `elements`
 
 Here I compare elements that are generated by numerical quadrature and 
 analytical integration. If one creates new elements, I suggest to start there 
 and perform similar comparisons before moving these comparisons to the tests.
 
-fdm
+--- 
+
+## `fdm`
 
 A couple of examples as comparison for FE. There is no plan at the moment, to 
 have extended fdm capabilities.
 
-fem 
+--- 
+
+## `fem` 
 
 Contains several application demonstration cases where topoptlab is (ab)used as 
 a FEM library. As TO first needs to solve the underlying physical problem, this is 
 the first step when applying TO to a new physical problem. 
 
-filter
+--- 
+
+## `filter`
 
 Here filters are compared to understand/illustrate what they do.
 
-linsolvers
+--- 
+
+## `linsolver`
 
 Demonstrates different solvers for linear systems. 
 
-optimization 
+--- 
+
+## `optimization`
 
 Contains simple demonstration of the (constrained) optimizers in topoptlab on
 simple cases like the Rosenbrock function to understand and compare performance.
 
-symfem
+--- 
+
+## `symfem`
 
 Here symfem is used to find analytical expressions for the element matrices. 
 This can be quite useful if one has an expression of the material properties 
@@ -38,7 +50,9 @@ as a function of microstructure parameters (e. g. laminates). The TO problem
 can then be formulated in terms of the microstructural parameters instead of 
 generic relative densities.
 
-topology optimization
+--- 
+
+## `topology optimization`
 
 Contains a number of example cases of topology optimization.
 
