+++ examples/symfem/README.md
@@ -1,63 +1,63 @@
-bodyforce.py
+### `bodyforce.py`
 
 Symbolically compute forces due to constant bodyforce.
 
-heatexpansion.py
+### `heatexpansion.py`
 
 Symbolically compute forces due to heat expansion.
 
-jacobian.py
+### `jacobian.py`
 
 Symbolically compute jacobian of (iso-)parametric map.
 
-laplacian.py 
+### `laplacian.py`
 
 Symbolically compute matrix of laplacian operator, e. g. for heat conduction, 
 diffusion, etc.
 
-lin_elasticity.py 
+### `lin_elasticity.py`
 
 Symbolically compute stiffness matrix for small deformation, small rotation 
 linear elasticity and forces arising due to uniform strain.
 
-mass_matrix.py 
+### `mass_matrix.py`
 
 Symbolically compute mass matrix.
 
-monomial.py 
+### `monomial.py`
 
 Symbolically compute matrix of a monomial of order >2.
 
-node_ordering.py
+### `node_ordering.py`
 
 Find index of each vertex by finding the element basis function that amounts 
 to 1.
 
-orthotropic_invariants.py
+### `orthotropic_invariants.py`
 
 Calculate stress invariants for orthotropic materials symbolically. Only viable 
 in 2D as sympy seems inefficient here.
 
-rotation_matrix_voigt.py
+### `rotation_matrix_voigt.py`
 
 Calculate the rotation matrix for a matrix in Voigt notation. 
 
-rotation_matrix.py
+### `rotation_matrix.py`
 
 Print rotation matrix.
 
-rotation_stiffness_tensor.py
+### `rotation_stiffness_tensor.py`
 
 Calculate rotation matrix for stiffness tensor.
 
-source.py
+### `source.py`
 
 Symbolically compute flux due to constant volume source.
 
-strain_measures.py
+### `strain_measures.py`
 
 Print different strain measures.
 
-volume_integral.py
+### `volume_integral.py`
 
 Symbolically compute volume integral of scalar field.
