+++ examples/optimization/README.md
@@ -1,8 +1,8 @@
-rosenbrock_bounded.py
+### `rosenbrock_bounded.py`
 
 Minimize the Rosenbrock function in the interval [-1.5,1.5].
 
-rosenbrock_constrained.py
+### `rosenbrock_constrained.py`
 
 Minimize the Rosenbrock function in the interval [-1.5,1.5] with constraint 
 that forces the sum of the squared variables to be equal to the number of 
