+++ examples/elements/README.md
@@ -1,8 +1,8 @@
-comparison.py
+### `comparison.py`
 
 Compares analytical and numerical integration for elements.
 
-consistency.py
+### `consistency.py`
 
 Checks that elements return consistent shapes depending on inputs.
 
