+++ examples/filter/README.md
@@ -1,14 +1,14 @@
-visualize_filter.py
+### `visualize_filter.py`
 
 Visualize effect of filter on a Dirac-delta peak, i. e.  a single filled 
 element.
 
-visualize_lengthscales.py
+### `visualize_lengthscales.py`
 
 visualize length scale violation, i. e. places in the design where details are
 smaller than a sphere of a given radius which resembles the desired lengthscale.
 
-volume_conservation.py
+### `volume_conservation.py`
 
 Short demonstration to check volume conservation after filtering.
 
