+++ examples/linsolvers/README.md
@@ -1,15 +1,18 @@
-anderson.py
+### `anderson.py
 
-laplacian.py
+Demonstrate convergence acceleration via Anderson acceleration. Taken from
+wikipedia and slightly adapted.
+
+### `laplacian.py`
 
 Demonstrate different solvers on finite difference laplacian.
 
-show_coarse.py
+### `show_coarse.py`
 
 Show coarse degrees of freedom for multigrid solver for a heatplate design or 
 blocks of block preconditioners.
 
-show_coarse_laplacian.py
+### `show_coarse_laplacian.py`
 
 Show coarse degrees of freedom for multigrid solver for finite difference 
 laplacian or blocks of block preconditioners.
