+++ examples/fdm/README.md
@@ -1,14 +1,14 @@
-allencahn_fdm.py
+### `allencahn_fdm.py`
 
 Solves Allen-Cahn equation via finite differences and explicit Euler. Only
 for demonstration and comparison with FEM.
 
-cahnhilliard_fdm.py
+### `cahnhilliard_fdm.py`
 
 Solves Cahn-Hilliard equation via finite differences and explicit Euler. Only
 for demonstration and comparison with FEM.
 
-cahnhilliard_pseudo.py
+### `cahnhilliard_pseudo.py`
 
 Solves Cahn-Hilliard equation via a pseudospectral method and explicit Euler. Only
 for demonstration and comparison with FEM.
