### 2025-11-17 4870301
Further expanded FE tutorial.
