+++ docs/source/introduction/crashcourse_fem.md
@@ -68,7 +68,7 @@ and von Neumann boundary conditions that fixes the value first derivative(s) to
 on a part of the boundary $\Gamma_N$ 
  
 ```{math}
-\nabla u\boldsymbol{n} = \boldsymbol{\partial u}_{N} \quad \text{on } \Gamma_{N}
+\nabla u = \boldsymbol{\partial u}_{N} \quad \text{on } \Gamma_{N}
 ```
 In solid mechanics, Dirichlet boundary conditions correspond to displacement 
 boundary conditions while in heat conduction or diffusion they correspond to 
@@ -77,7 +77,7 @@ are often formulated differently, as in real cases rarely the gradient of the
 state variable $u$ is available, but instead the **flux** is measured, so the
 above equation is simply rescaled by some material constants 
 ```{math}
-\boldsymbol{K} \nabla u) = q_{N} \quad \text{on } \Gamma_{N}
+\boldsymbol{K} \nabla u = q_{N} \quad \text{on } \Gamma_{N}
 ```
 which however does not change the nature of the boundary condition: it is still
 a boundary condition in terms of the first order derivatives.
@@ -87,7 +87,7 @@ conditions by a **weight function** $w$ and integrate over the domain $\Omega$,
 so we end up with
 ```{math}
 \int_\Omega w \, (\mathcal{L}(u) - f) \, dV = 0. \\ 
-\int_{\Gamma_{N}} w (\nabla u)^T \boldsymbol{n} dA  = u_{N} 
+\int_{\Gamma_{N}} w ( \boldsymbol{K} \nabla u)^T \boldsymbol{n} dA  = u_{N} 
 ```
 The weak form is fully equivalent to the strong form only if $u$ and $w$ 
 satisfy some conditions. Among these conditions are $w!=0$ and that the weak 
@@ -112,7 +112,7 @@ weak form via standard procedure i) multiply PDE and boundary conditions by $w$
 ii) integrate over domain $\Omega$:
 ```{math}
 \int_\Omega w \nabla \cdot (\boldsymbol{K} \nabla \phi) dV = \int_\Omega w f dV \\
-\int_{\Gamma_{N}} \left(w \boldsymbol{K} \nabla \phi\right) \cdot \boldsymbol{n} dA
+\int_{\Gamma_{N}} w\left(\boldsymbol{K} \nabla \phi\right)^T \boldsymbol{n} dA
 ```
 Technically one can stop now as this is a correct weak from, but we want to 
 reduce the highest order derivative as much as we can as in FEM this means we 
@@ -132,7 +132,7 @@ w\nabla \boldsymbol{v} = \nabla w \cdot \boldsymbol{v} - \nabla \cdot (w\boldsym
 ```
 We then insert $\boldsymbol{v}=\nabla \phi$ and rewrite the left hand side to
 ```{math}
-\int_\Omega w \nabla \cdot (\boldsymbol{K} \nabla \phi) dV = \int_\Omega \nabla w \cdot \boldsymbol{K} \nabla \phi dV - \int_\Omega \nabla \cdot \left(w \boldsymbol{K} \nabla \phi\right) dV
+\int_\Omega w \nabla \cdot (\boldsymbol{K} \nabla \phi) dV = \int_\Omega \nabla w \cdot \boldsymbol{K} \nabla \phi dV - \int_\Omega \nabla \cdot w\left(\boldsymbol{K} \nabla \phi\right) dV
 ```
 If we inspect this closer, we recognize that the second term on the right hand 
 side is the volume integral of the divergence of the flow 
@@ -147,14 +147,14 @@ expresses
[…]
