### 2025-11-17 cbcf1e0
Add reference to paper.
