### 2025-11-18 f9c6012
Added explicit link to README and examples into paper.
