### 2025-11-19 c494a61
added scipy iterative solvers for asymmetric matrices and added new block creation routines with tests.
