+++ tests/test_solver.py
@@ -9,6 +9,8 @@ from topoptlab.example_bc.lin_elast import mbb_2d,mbb_3d
 @pytest.mark.parametrize('nelx, nely, volfrac, ft, rmin, solver, preconditioner, bcs',
                          [(10,3,0.5,0,2.4,"scipy-direct",None,mbb_2d),
                           (10,3,0.5,1,2.4,"scipy-cg",None,mbb_2d),
+                          (10,3,0.5,1,2.4,"scipy-bicg",None,mbb_2d),
+                          (10,3,0.5,1,2.4,"scipy-gmres","scipy-ilu",mbb_2d),
                           (10,3,0.5,1,2.4,"scipy-cg","scipy-ilu",mbb_2d),
                           (10,3,0.5,1,2.4,"cvxopt-cholmod",None,mbb_2d),
                           (10,3,0.5,1,2.4,"scipy-cg","pyamg-adaptive_sa",mbb_2d),
@@ -41,7 +43,8 @@ def test_compliance_1(nelx, nely, volfrac,
                   rmin=rmin, ft=ft,
                   filter_mode="matrix",optimizer="oc",
                   bcs=bcs,
-                  lin_solver_kw = {"name": solver},
+                  lin_solver_kw = {"name": solver,
+                                   "maxiter": 1000},
                   preconditioner_kw = {"name": preconditioner},
                   output_kw = {"file": None,
                                "display": False,

[…]
