### 2025-11-20 37b03af
add all scipy sparse iterative solvers and tests. Also re-use factoirizations which was forogt before. Also eliminate unecessary print statements. Further add logging to iterative linear solvers.
