+++ examples/topology_optimization/compliant_mechanisms/force_inverter3d.py
@@ -53,7 +53,8 @@ if __name__ == "__main__":
          assembly_mode="lower",
          bcs=partial(forceinverter_3d,fixation_mode="line") , 
          obj_func=var_maximization ,obj_kw={"l": l},
-         ft=ft, filter_mode="matrix",optimizer="mma",
+         ft=ft, filter_mode="matrix",
+         optimizer="mma", optimizer_kw={"move": 0.05},
          output_kw = {"file": "force-inverter_3d",
                       "display": display,
                       "export": export,
