### 2025-11-24 363d627
Added nonlinearity to symbolic laplacian/heat conductivity. Need to still test. Also implemented interval element for numerical quadrature, but need tests.
