+++ topoptlab/elements/linear_interval.py
@@ -1,0 +1,327 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+from typing import Tuple,Any,Union,List
+from warnings import warn
+
+import numpy as np
+
+def create_edofMat(nelx: int, 
+                   nnode_dof: int,
+                   dtype: type = np.int32, 
+                   **kwargs: Any
+                   ) -> Tuple[np.ndarray,None,None,None,None]:
+    """
+    Create element degree of freedom matrix for linear Lagrangian elements in
+    a regular mesh.
+
+    Parameters
+    ----------
+    nelx : int
+        number of elements in x direction.
+    nnode_dof : int
+        number of node degrees of freedom.
+
+    Returns
+    -------
+    edofMat : np.ndarray
+        element degree of freedom matrix
+    n1 : None
+        purely there for compatibility with 2/3 dimensions.
+    n2 : None
+        purely there for compatibility with 2/3 dimensions.
+    n3 :  None
+        purely there for compatibility with 3 dimensions.
+    n4 :  None
+        purely there for compatibility with 3 dimensions.
+    """
+    # create arrays for indexing
+    elx = np.arange(nelx, dtype=dtype)[:,None]
+    #
+    edofMat = np.column_stack((elx,elx+1))*nnode_dof
+    edofMat = np.repeat(edofMat,nnode_dof,axis=1)
+    edofMat = edofMat + np.tile(np.arange(nnode_dof,dtype=dtype),2)[None,:]
+    return edofMat, None, None, None, None
+
+def apply_pbc(edofMat: np.ndarray, 
+              pbc: Union[List,np.ndarray],
+              nelx: int, 
+              nnode_dof: int,
+              dtype: type = np.int32, 
+              **kwargs: Any) -> np.ndarray:
+    """
+    Convert a given element-degree-of-freedom matrix (edofMat) of a regular
+    mesh of first order Lagrangian interval with free boundary conditions to 
+    the corresponding edofMat with periodic boundary conditions by re-labelling 
+    the nodal degrees of freedom
+
+    Parameters
+    ----------
+    edofMat : np.ndarray of shape (nel,2*nnode_dof)
+        element-degree-of-freedom matrix (edofMat) of a regular
+        mesh of first order Lagrangian quadrilateral elements with free
+        boundary conditions.
+    pbc : list or np.ndarray of shape/len 1 with elements of datatype bool
+        periodic boundary condition flags.
+    nelx : int
+        number of elements in x-direction.
+    nnode_dof : int
+        number of degrees of freedom per node.
+    dtype : type
+        datatype of element degrees of freedom. Should be just large enough to
+        store the highest number of the degrees of freedom to save memory. For
+        practical purposes np.int32 should do the job.
+
+    Returns
+    -------
+    edofMat_new : np.ndarray of shape (nel,2*nnode_dof)
+        element-degree-of-freedom matrix (edofMat) of a regular
+        mesh of first order Lagrangian interval elements with periodic
+        boundary conditions with re-labelled nodal degrees of freedom.
+    """
+    if pbc[0]:
+        # reassign indices
+        edofMat[-1,-nnode_dof] = edofMat[0,:nnode_dof]
+    return edofMat
+
+def check_inputs(xi: Union[float,np.ndarray],
+                 xe: Union[None,np.ndarray] = None,
+                 all_elems: bool = False,
+                 **kwargs: Any):
+    """
+    Check coordinates and provided element node information to be consistent.
+    If necessary transform inputs to make them consistent.
+
+    Parameters
+    ----------
+    xi : float or np.ndarray
[…]
