### 2025-11-28 d71eab9
added 1D element and further steps towards nonlinearity for symbolic elements.
