### 2025-12-05 57d7bd8
More tutorials, but got crashcourse not entirely finished.
