+++ docs/source/introduction.rst
@@ -9,4 +9,5 @@ Introduction and Tutorials to Topology Optimization
    introduction/adjoint_analysis
    introduction/meshing
    introduction/crashcourse_fem
+   introduction/variable-thickness-sheet
    introduction/references
