### 2025-12-08 b603bdc
small changes to solver test. cgs appears not to be stable on the compliance test.
