+++ tests/test_solver.py
@@ -9,7 +9,7 @@ from topoptlab.example_bc.lin_elast import mbb_2d,mbb_3d
                          [(10,3,0.5,0,2.4,"scipy-direct",None,mbb_2d),
                           (10,3,0.5,1,2.4,"scipy-cg",None,mbb_2d),
                           (10,3,0.5,1,2.4,"scipy-bicg",None,mbb_2d),
-                          (10,3,0.5,1,2.4,"scipy-cgs",None,mbb_2d),
+                          (10,3,0.5,1,2.4,"scipy-cgs","scipy-ilu",mbb_2d),
                           (10,3,0.5,1,2.4,"scipy-minres",None,mbb_2d),
                           (10,3,0.5,1,2.4,"scipy-lgmres",None,mbb_2d),
                           (10,3,0.5,1,2.4,"scipy-gmres","scipy-ilu",mbb_2d),
