### 2025-12-12 a1387b6
centre some maths
