### 2025-12-12 979aad9
0 or 1 (as 0/1 looks like a fraction)
