+++ paper/paper.md
@@ -45,7 +45,7 @@ via simple relationships as in the popular modified SIMP approach [@sigmund2007m
 
 $$A(x_e) = A_{min} + (A_{0}-A_{min})x_{e}^{k}\text{ with }0\leq x_e\leq 1$$
 
-where $k$ is a penalization factor ensuring densities close to 0/1, $A_0$ 
+where $k$ is a penalization factor ensuring densities close to 0 or 1, $A_0$ 
 the property of the full material and $A_{min}$ a small value to prevent 
 singularities in the physical problem. The final design then emerges from the 
 optimal density distribution.  
