### 2025-12-12 d7bbbbb
some minor edits in paper
