### 2025-12-15 394aeee
First naive implementation of Reuss-Voigt. Need to put in stable inversion.
