### 2025-12-15 2b828d5
Added more stable options, but need to think about more stable version for reuss.
