### 2025-12-16 9b1056f
small changes to Reuss bound
