### 2025-12-16 a9db6fa
Added functions for node ordering.
