### 2026-01-04 96408d5
fixed some objectives and generalized the TO main routine.
