+++ examples/topology_optimization/compliance_minimization/heatplate2d.py
@@ -36,7 +36,9 @@ if __name__ == "__main__":
     if len(sys.argv)>9:
         write_log = bool(int(sys.argv[9]))
     #
-    main(nelx=nelx, nely=nely, volfrac=volfrac, penal=penal, rmin=rmin, 
+    main(nelx=nelx, nely=nely, volfrac=volfrac, 
+         matinterpol_kw={"eps":1e-9, "penal": penal},
+         rmin=rmin, 
          ft=ft, filter_mode="matrix", optimizer="oc",nouteriter=1000,
          #lin_solver_kw = {"name": "cvxopt-cholmod"},
          #lin_solver_kw = {"name": "topoptlab-cg"}, preconditioner_kw = {"name": "pyamg-pyamg-ruge_stuben"},
