### 2026-01-08 dfde194
Added typing to objectives.
