+++ topoptlab/topology_optimization.py
@@ -10,6 +10,7 @@ from scipy.ndimage import convolve
 from matplotlib.colors import Normalize
 import matplotlib.pyplot as plt
 # functions to create filters
+from topoptlab.filter.filter import TOFilter 
 from topoptlab.filter.convolution_filter import assemble_convolution_filter
 from topoptlab.filter.helmholtz_filter import assemble_helmholtz_filter
 from topoptlab.filter.matrix_filter import assemble_matrix_filter
@@ -47,6 +48,7 @@ def main(nelx: int, nely: int,
          volfrac: float, #penal: float, 
          rmin: float, 
          ft: int = 1,
+         filter_kw: Dict = {},
          simulation_kw: Dict = {"grid": "regular",
                                 "element order": 1,
                                 "meshfile": None},
@@ -214,8 +216,12 @@ def main(nelx: int, nely: int,
     x = volfrac * np.ones( (n,1), dtype=float,order='F')
     #x = np.random.rand(n)
     #x = x/x.mean() * volfrac
-    xTilde = x.copy()
     xPhys = x.copy()
+    # intermediate filter variables
+    if isinstance(ft, list):
+        xTilde = dict()
+        for i in range(len(ft)):
+            xTilde.append(x.copy)
     # initialize arrays for gradients
     dobj = np.zeros( x.shape,order="F")
     # initialize constraints
@@ -254,8 +260,6 @@ def main(nelx: int, nely: int,
             xPhys[mask] = 1.
     else:
         raise ValueError("Unknown optimizer: ", optimizer)
-    # Young's modulus 
-    E = materials_kw["E"]
     # get element stiffness matrix
     KE = lk(l=l)
     # infer nodal degrees of freedom assuming that we have 4/8 nodes in 2/3 D
@@ -332,7 +336,12 @@ def main(nelx: int, nely: int,
         mapping = partial(map_eltovoxel,
                           nelx=nelx,nely=nely,nelz=nelz)
     # prepare functions to invert this mapping if we use the convolution filter
-    if filter_mode == "convolution" and ndim == 2:
+    if isinstance(ft, TOFilter):
+        ft = [ft(nelx=nelx,nely=nely,nelz=nelz,rmin=rmin,**filter_kw)]
+    elif isinstance(ft, list):
+        ft = [ft_obj(nelx=nelx,nely=nely,nelz=nelz,rmin=rmin,**filter_kw) \
+              for ft_obj in ft]
+    elif filter_mode == "convolution" and ndim == 2:
         invmapping = partial(map_imgtoel,
                              nelx=nelx,nely=nely)
     elif filter_mode == "convolution" and ndim == 3:
@@ -510,7 +519,27 @@ def main(nelx: int, nely: int,
         log.debug("[DEBUG] Pre-Sensitivity Filter: it.: {0}, dobj: {1:.10f}, dv: {2:.10f}".format(
                   loop, np.max(dobj), np.min(dconstrs)))
         # Sensitivity filtering:
-        if ft == 0 and filter_mode == "matrix":
+        if isinstance(ft, list):
+            dobj[:] = ft[-1].apply_filter_dx(x_filtered=xPhys,
+                                    dx_filtered=dobj)
+            dconstrs[:] = ft[-1].apply_filter_dx(x_filtered=xPhys,
+                                                 dx_filtered=dconstrs)
+            if len(ft) > 1:
+                for i in range(len(ft)-1,-1,-1):
+                    dobj[:] = ft[-1].apply_filter_dx(x_filtered=xTilde[i],
+                                                     dx_filtered=dobj)
[…]
