### 2026-01-12 c368d82
implemented basic symbolic hyperelasticity. now need to finish nonlinear elastic element definition.
