### 2026-01-12 22a8b31
Exchange of curly quotation marks in paper.
