+++ paper/paper.md
@@ -13,7 +13,7 @@ affiliations:
    index: 1
  - name: CNR – Consiglio Nazionale delle Ricerche, Istituto di Chimica della Materia Condensata e di Tecnologie per l’Energia, Via R. Cozzi 53, 20125 Milano, Italy
    index: 2 
-date: 25 August 2025
+date: 13 January 2026
 bibliography: paper.bib
 ---
 
