+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.0.1
+current_version = 1.0.0
 tag = True
 commit = True
 verbose = True
