### 2026-01-13 #9105 @stefanhiemer
@gkthiruvathukal the archived Zenodo DOI is https://doi.org/10.5281/zenodo.18231035 and the release tag is `1.0.0`. I hope I did not miss anything. Do you need anything else from me?
