+++ README.md
@@ -1,6 +1,6 @@
 # topoptlab 
 [![Documentation Status](https://readthedocs.org/projects/topoptlab/badge/?version=latest)](https://topoptlab.readthedocs.io/en/latest/?badge=latest)
-[![PyPI](https://img.shields.io/pypi/v/topoptlab?color=blue&label=PyPI&logo=pypi&logoColor=white)](https://pypi.org/project/topoptlab/)
+[![PyPI](https://img.shields.io/pypi/v/topoptlab)](https://pypi.org/project/topoptlab/)
 [![CI](https://github.com/stefanhiemer/topoptlab/actions/workflows/ci.yaml/badge.svg?branch=main)](https://github.com/stefanhiemer/topoptlab/actions/workflows/ci.yaml)
 
 This project provides a collection of topology optimization techniques, many of 
