### 2026-01-16 e7d6645
added a matrix utility routines and first steps towards hyperelasticity. still need to write tests there and confirm with textbooks. also might have to parallelize some routines for simplification.
