+++ topoptlab/symbolic/nonlin_elasticity.py
@@ -1,29 +1,52 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
-from typing import Union
-from math import floor
+from typing import Callable, Dict, List, Union 
 
 from symfem.symbols import x
 from symfem.functions import MatrixFunction
 
 from topoptlab.symbolic.cell import base_cell
 from topoptlab.symbolic.parametric_map import jacobian
-from topoptlab.symbolic.matrix_utils import eye, simplify_matrix, generate_constMatrix
-from topoptlab.symbolic.strain_measures import cauchy_strain,dispgrad_matrix 
+from topoptlab.symbolic.matrix_utils import eye, simplify_matrix,\
+                                            inverse, generate_constMatrix,\
+                                            generate_FunctMatrix, to_voigt,\
+                                            from_voigt, to_column
+from topoptlab.symbolic.strain_measures import cauchy_strain,dispgrad_matrix, def_grad
+from topoptlab.symbolic.stress_conversions import cauchy_to_pk1, pk2_to_pk1
+from topoptlab.symbolic.hyperelasticity import stvenant_2pk, stvenant_cauchy
 
 def residual(ndim : int,
-             s : Union[None,MatrixFunction],
+             sigma : Union[None,Callable],
+             pk1: Union[None,Callable] = None,
+             pk2: Union[None,Callable] = None,
+             material_params: Dict = dict(),
              element_type : str ="Lagrange",
-             order : int = 1) -> MatrixFunction:
+             order : int = 1,
+             return_symbols: bool =False) -> MatrixFunction:
     """
-    Symbolically compute the residual for nonlinear 
-    elasticity formulated in terms of the 1. PK stress P.
+    Symbolically compute the residual for nonlinear elasticity formulated in 
+    terms of the 1. Piola-Kirchhoff (PK) stress P and the deformation gradient
+    F.
+    
+    The constitutive behaviouris given by a stress relationship supplied via a 
+    Callable of the stress. You must decide in which stress the material law is 
+    formulated and depending on that the solver converts the given callable to 
+    calculate the stress to the 1. PK. The callable should calculate the stress
+    based on the deformation gradient F at the current iteration with nodal 
+    displacements u0.
 
     Parameters
     ----------
     ndim : int
         number of spatial dimensions. Must be between 1 and 3.
-    s : None or symfem.functions.MatrixFunction
-        stiffness tensor .
+    sigma : None or callable
+        returns Cauchy stress in Voigt form. This is what an Elmer UMAT should 
+        do.
+    pk1 : None or callable
+        returns 1. PK stress in flattened shape. 
+    pk2 : None or callable
+        returns 2. PK stress in Voigt form. 
+    material_params : Dict,
+        dictionary containing material parameters.
     element_type : str
         type of element.
     order : int
@@ -32,33 +55,88 @@ def residual(ndim : int,
     Returns
     -------
     stiffness_matrix : symfem.functions.MatrixFunction
-        symbolic stiffness matrix.
+        symbolic residual of shape (n_nodes,1).
 
     """
     #
     vertices, nd_inds, ref, basis  = base_cell(ndim)
-    # anisotropic stiffness tensor or equivalent in Voigt notation 
-    if s is None:
-        s = generate_constMatrix(ncol=int((ndim**2 + ndim) /2),
-                                 nrow=int((ndim**2 + ndim) /2),
-                                 name="s")
     #
-    u0 = generate_constMatrix(ncol=1, nrow=len(basis), name="u0")
+    u,u_symbs = generate_constMatrix(ncol=1, nrow=len(basis)*ndim, name="u",
+                                     return_symbols=True)
     # calculate deformation gradient
-    I = eye(size=ndim)
[…]
