+++ topoptlab/elements/strain_measures.py
@@ -12,10 +12,10 @@ def eng_strain(xi: np.ndarray, eta: np.ndarray,
                all_elems: bool = False, return_detJ: bool = False,
                **kwargs: Any):
     """
-    Engineering strain via the B matrix for bilinear quadrilateral Lagrangian 
-    element to calculate to calculate strains, stresses etc. from nodal values:
-    
-    eps = B@u
+    Return the the B matrix to calculate the infinitesimal or engineering 
+    strain via from nodal displacements u:
+        
+        eps = B@u
     
     Parameters
     ----------
@@ -38,7 +38,8 @@ def eng_strain(xi: np.ndarray, eta: np.ndarray,
         
     Returns
     -------
-    B : np.ndarray, shape (ncoords,3,8) or (nels,3,8)
+    B : np.ndarray, shape (ncoords,(ndim**2 + ndim)/2,n_nodes*ndim) or 
+        (nels,(ndim**2 + ndim)/2,n_nodes*ndim)
         B matrix.
     detJ : np.ndarray, shape (ncoords) or (nels)
            determinant of Jacobian.
@@ -72,6 +73,74 @@ def eng_strain(xi: np.ndarray, eta: np.ndarray,
         B[ndim+k][j::ndim] = gradN[:,:,i]
         #
         i,j = (i+1)%ndim , (j+1)%ndim
+    if not return_detJ:
+        return B
+    else:
+        return B, detJ
+    
+def disp_gradient(xi: np.ndarray, eta: np.ndarray, 
+                  xe: np.ndarray,
+                  invjacobian: np.ndarray, 
+                  shape_functions_dxi: np.ndarray,
+                  zeta: Union[None,np.ndarray] = None, 
+                  check_fnc: Union[None,Callable] = None,
+                  all_elems: bool = False, return_detJ: bool = False,
+                  **kwargs: Any):
+    """
+    Return the the B matrix to calculate the flattened displacement gradient h
+    ('C' ordering)from nodal displacements u:
+        
+        h = B_h@u
+    
+    Parameters
+    ----------
+    xi : np.ndarray
+        x coordinate in the reference domain of shape (ncoords).
+    eta : np.ndarray
+        y coordinate in the reference domain of shape (ncoords). Coordinates are assumed to be
+        in the reference domain.
+    xe : np.ndarray
+        coordinates of element nodes shape (nels,n_nodes,ndim). nels must be 
+        either 1, ncoords/4 or the same as ncoords. The two exceptions are if 
+        ncoords = 1 or all_elems is True. 
+        Please look at the definition/function of the shape function, then the 
+        node ordering is clear.
+    all_elems : bool
+        if True, coordinates are evaluated for all elements. Useful for 
+        creating elements etc.
+    return_detJ : bool
+        if True, return determinant of jacobian.
+        
+    Returns
+    -------
+    B_h : np.ndarray, shape (ncoords,ndim**2,n_nodes*ndim) or 
+          (ndim**2,n_nodes*ndim)
+        matrix for calculating the displacement gradient matrix.
+    detJ : np.ndarray, shape (ncoords) or (nels)
+           determinant of Jacobian.
+        
+    """
+    #
+    nel, n_nodes, ndim = xe.shape
+    # check coordinates and node data for consistency
+    if hasattr(check_fnc, '__call__') and zeta is None:
+        xe,xi,eta,zeta = check_fnc(xi=xi,eta=eta,xe=xe,
+                                   all_elems=all_elems) 
+    # collect inverse jacobians
+    if not return_detJ:
+        invJ = invjacobian(xi=xi,eta=eta,zeta=zeta,xe=xe,
[…]
