+++ tests/symbolic/test_hyperelasticity.py
@@ -2,7 +2,7 @@ from sympy import symbols
 
 import pytest
 
-from topoptlab.symbolic.hyperelasticity import calculate_2PK, stvenant_engdensity, stvenant_2PK
+from topoptlab.symbolic.hyperelasticity import calculate_2pk, stvenant_engdensity, stvenant_2pk
 
 @pytest.mark.parametrize('ndim',
                          [(1) ])
@@ -11,5 +11,5 @@ def test_stvenant_engdensity(ndim):
     #
     eng=stvenant_engdensity(ndim=ndim)
     #
-    assert calculate_2PK(eng_density=eng,E_v=None,ndim=ndim)==stvenant_2PK(E_v=None,c=None,ndim=ndim)
+    assert calculate_2pk(eng_density=eng,E=None,ndim=ndim)==stvenant_2pk(E=None,c=None,ndim=ndim)
     return 

[…]
