### 2026-01-19 67ed779
Neo-hookean symbolics of huhu regularization is done. Hessian still needs to be done completely and elements finalized.
