+++ topoptlab/symbolic/tensor_invariants.py
@@ -8,42 +8,10 @@ from symfem.functions import ScalarFunction,MatrixFunction
 
 from topoptlab.symbolic.matrix_utils import check_square, diag, trace
 
-def main_invariants(A: MatrixFunction) -> Tuple:
-    """
-    Return main tensor invariants J of A. Does not work if dim(A) > 3.
-
-    Parameters
-    ----------
-    A : symfem.functions.MatrixFunction
-        square matrix.
-
-    Returns
-    -------
-    J1 : symfem.functions.ScalarFunction
-        first main invariant
-    J2 : symfem.functions.ScalarFunction
-        second main invariant
-    J3 : symfem.functions.ScalarFunction
-        third main invariant
-    """
-    #
-    check_square(A)
-    #
-    ndim = A.shape[0]
-    #
-    if ndim == 1:
-        return A
-    elif ndim == 2:
-        return trace(A),A.det()
-    elif ndim == 3:
-        I1,I2,I3 = principal_invariants
-        return I1, I1**2 - 2*I2, I1**3 - 3*(I1*I2 + I3)
-    else:
-        raise ValueError("This function only works for square matrices ndim < 4.")
-
 def principal_invariants(A: MatrixFunction) -> Tuple:
     """
-    Return principal tensor invariants I of A. Does not work if dim(A) > 3.
+    Return the principal tensor invariants `I` of MatrixFunction `A`. Works 
+    only for dim(A) <= 3.
 
     Parameters
     ----------
@@ -53,23 +21,59 @@ def principal_invariants(A: MatrixFunction) -> Tuple:
     Returns
     -------
     I1 : symfem.functions.ScalarFunction
-        first principal invariant
+        first principal invariant (trace).
     I2 : symfem.functions.ScalarFunction
-        second principal invariant
+        second principal invariant.
     I3 : symfem.functions.ScalarFunction
-        third principal invariant
+        third principal invariant (determinant).
     """
-    #
     check_square(A)
-    #
     ndim = A.shape[0]
-    #
+
     if ndim == 1:
-        return A
+        return A[0, 0]
     elif ndim == 2:
-        return trace(A),A.det()
+        return trace(A), A.det()
     elif ndim == 3:
-        return trace(A), diag([1/2,1/2,1/2])@(trace(A)**2 - trace(A@A)), A.det()
+        I1 = trace(A)
+        I2 = (I1**2 - trace(A @ A)) / 2
+        I3 = A.det()
+        return I1, I2, I3
     else:
-        raise ValueError("This function only works for square matrices ndim < 4.")
-    
+        raise ValueError("This function only works for square matrices with dim ≤ 3.")
+
+
+def main_invariants(A: MatrixFunction) -> Tuple:
+    """
+    Return the main tensor invariants `I` of MatrixFunction `A` expressed in 
+    terms of the principal invariants. Works only for dim(A) <= 3.
+
+    Parameters
+    ----------
+    A : symfem.functions.MatrixFunction
+        square matrix.
+
+    Returns
+    -------
+    J1 : symfem.functions.ScalarFunction
+        first main invariant.
+    J2 : symfem.functions.ScalarFunction
+        second main invariant.
+    J3 : symfem.functions.ScalarFunction
+        third main invariant.
+    """
+    check_square(A)
+    ndim = A.shape[0]
+
+    if ndim == 1:
+        return A[0, 0]
+    elif ndim == 2:
+        return trace(A), A.det()
+    elif ndim == 3:
+        I1, I2, I3 = principal_invariants(A)
+        J1 = I1
+        J2 = I1**2 - 2*I2
+        J3 = I1**3 - 3*I1*I2 + 3*I3
+        return J1, J2, J3
+    else:
+        raise ValueError("This function only works for square matrices with dim ≤ 3.")
