### 2026-01-20 cc99696
Added tests towards hyperelasticity and HuHu. Also some clean-up and generalizations.
