+++ tests/elements/test_mass.py
@@ -37,10 +37,10 @@ def test_compareanalyt(xe):
 @pytest.mark.parametrize('xe',
                          [(array([[[-1,-1],[1,-1],[1,1],[-1,1]],
                                   [[-2,-2.1],[2.1,-2],[2,2],[-2,2]]])),
-                          array([[[-1,-1,-1],[1,-1,-1],[1,1,-1],[-1,1,-1],
+                          (array([[[-1,-1,-1],[1,-1,-1],[1,1,-1],[-1,1,-1],
                                   [-1,-1,1],[1,-1,1],[1,1,1],[-1,1,1]],
                                  [[-2.1,-2,-2],[2.1,-2,-2],[2,2.1,-2],[-2,2,-2],
-                                  [-2,-2,2],[2,-2,2.1],[2,2,2],[-2,2.1,2]]])])
+                                  [-2,-2,2],[2,-2,2.1],[2,2,2],[-2,2.1,2]]]))])
 
 def test_consist(xe):
     
