### 2026-01-21 34461de
Numerical integration for HuHu might work now. Still need to test and compare.
