+++ topoptlab/elements/huhu_2d.py
@@ -3,17 +3,20 @@ from typing import Any
 
 import numpy as np
 
-from topoptlab.elements.strain_measures import disp_gradient
-from topoptlab.elements.bilinear_quadrilateral import shape_functions_hessian,invjacobian
+from topoptlab.elements.strain_measures import dispgrad_matrix
+from topoptlab.elements.bilinear_quadrilateral import shape_functions_dxi,\
+                                                      shape_functions_hessian,\
+                                                      invjacobian
 from topoptlab.fem import get_integrpoints
 
 def _lk_huhu_2d(xe: np.ndarray, 
                 ue: np.ndarray,
-                a: np.ndarray,
-                kr: np.ndarray,     
+                exponent: np.ndarray,
+                kr: np.ndarray,
+                mode="newton",
                 quadr_method: str = "gauss-legendre",
                 t: np.ndarray = np.array([1.]),
-                nquad: int = 2,
+                nquad: int = 4,
                 **kwargs: Any) -> np.ndarray:
     """
     Create element stiffness matrix for 2D HuHu regularization with
@@ -32,9 +35,9 @@ def _lk_huhu_2d(xe: np.ndarray,
         clear.
     ue : np.ndarray,shape (nels,8).
         nodal displacements.
-    a : np.ndarray, shape (nels,1) or
+    exponent : np.ndarray, shape (nels) or float
         exponent.
-    kr : np.ndarray, shape (nels,1) or
+    kr : np.ndarray, shape (nels) or float
         regularization strength.
     quadr_method: str or callable
         name of quadrature method or function/callable that returns coordinates of
@@ -58,11 +61,11 @@ def _lk_huhu_2d(xe: np.ndarray,
         xe = xe[None,:,:]
     nel = xe.shape[0]
     #
-    if len(a.shape) == 1:
-        a = a[None,:]
+    if isinstance(exponent,float):
+        exponent = np.array([exponent])
     #
-    if len(kr.shape) == 1:
-        kr = kr[None,:]
+    if isinstance(kr,float):
+        kr = np.array([kr])
     #
     if isinstance(t,float):
         t = np.array([t])
@@ -77,32 +80,61 @@ def _lk_huhu_2d(xe: np.ndarray,
     Jinv = Jinv.reshape(nel,nq,2,2)
     detJ = detJ.reshape(nel,nq)
     # collect hessian in ref. space
-    hessian = shape_functions_hessian(xi=xi, eta=eta) # (nq,n_basis,2,2)
+    B_hessian = shape_functions_hessian(xi=xi, eta=eta) # (nq,n_basis,2,2)
     # apply isop. map
-    hessian = Jinv.transpose((0,1,3,2))[:,:,None,:,:]@hessian[None,:,:,:]@\
-              Jinv[:,:,None,:,:]#.transpose((0,1,3,2))
+    B_hessian = Jinv.transpose((0,1,3,2))[:,:,None,:,:]@B_hessian[None,:,:,:]@\
+                Jinv[:,:,None,:,:]#.transpose((0,1,3,2))
+    #print(hessian.shape)
     # flatten hessian
-    hessian = hessian.reshape((nel,nq,ndim**2))
-    import sys 
-    sys.exit()
+    B_hessian = B_hessian.reshape(B_hessian.shape[:3]+tuple([ndim**2]))
+    B_hessian = B_hessian.transpose( (0,1,3,2) )
+    #print(hessian.shape)
+    #print(hessian[0,0])
+    B_hessian = np.kron(B_hessian,np.eye(ndim))
+    #print(hessian[0,0])
+    #print("B_hessian",B_hessian.shape)
[…]
