+++ tests/elements/test_strain_measures_numerical.py
@@ -1,17 +1,15 @@
-from numpy import array,stack,vstack
+from numpy import array,eye,stack,tile,vstack
 from numpy.testing import assert_allclose
 
 import pytest
 
 
-from topoptlab.elements.strain_measures import eng_strain  
+from topoptlab.elements.strain_measures import infini_strain_matrix  
 
 from topoptlab.elements.bilinear_quadrilateral import invjacobian as \
                                                       invjacobian_bilin
 from topoptlab.elements.bilinear_quadrilateral import shape_functions_dxi as \
                                                       shape_functions_dxi_bilin
-from topoptlab.elements.bilinear_quadrilateral import check_inputs as \
-                                                      check_inputs_bilin
 from topoptlab.elements.bilinear_quadrilateral import bmatrix as \
                                                       bmatrix_bilin
 
@@ -24,10 +22,10 @@ def test_consistency_bilinear(xe,xi,eta):
     
     
     
-    bmat_general = eng_strain(eta=eta,xi=xi,xe=xe, all_elems=False,
-                              invjacobian=invjacobian_bilin,
-                              shape_functions_dxi=shape_functions_dxi_bilin,
-                              check_fnc=check_inputs_bilin)
+    bmat_general = infini_strain_matrix(eta=eta,xi=xi,zeta=None,
+                                        xe=xe, all_elems=False,
+                                        invjacobian=invjacobian_bilin,
+                                        shape_functions_dxi=shape_functions_dxi_bilin)
     #
     bmat_quadri = bmatrix_bilin(eta=eta,xi=xi,xe=xe)
     #
@@ -39,8 +37,6 @@ from topoptlab.elements.trilinear_hexahedron import invjacobian as \
                                                     invjacobian_trilin
 from topoptlab.elements.trilinear_hexahedron import shape_functions_dxi as \
                                                     shape_functions_dxi_trilin
-from topoptlab.elements.trilinear_hexahedron import check_inputs as \
-                                                    check_inputs_trilin
 from topoptlab.elements.trilinear_hexahedron import bmatrix as \
                                                     bmatrix_trilin
 
@@ -53,11 +49,10 @@ from topoptlab.elements.trilinear_hexahedron import bmatrix as \
 
 def test_consistency_trilinear(xe,xi,eta,zeta): 
     
-    bmat_general = eng_strain(eta=eta,xi=xi,xe=xe, all_elems=False,
+    bmat_general = infini_strain_matrix(eta=eta,xi=xi,xe=xe, all_elems=False,
                               zeta=zeta,
                               invjacobian=invjacobian_trilin,
-                              shape_functions_dxi=shape_functions_dxi_trilin,
-                              check_fnc=check_inputs_trilin)
+                              shape_functions_dxi=shape_functions_dxi_trilin)
     #
     bmat_hex = bmatrix_trilin(eta=eta,xi=xi,zeta=zeta,xe=xe)
     #
@@ -119,23 +114,21 @@ def test_consistency_trilinear(xe,xi,eta,zeta):
 def test_eng_strain(xe,xi,eta,zeta,u,eps):
     
     if zeta is None:
-        bmat = eng_strain(eta=eta,xi=xi,xe=xe, all_elems=False,
-                          zeta=zeta,
-                          invjacobian=invjacobian_bilin,
-                          shape_functions_dxi=shape_functions_dxi_bilin,
-                          check_fnc=check_inputs_bilin)
+        bmat = infini_strain_matrix(eta=eta,xi=xi,zeta=zeta, 
+                                    xe=xe, all_elems=False,
+                                    invjacobian=invjacobian_bilin,
+                                    shape_functions_dxi=shape_functions_dxi_bilin)
     else:
[…]
