### 2026-01-22 56e047a
Huhu compares well with sigmund, but Newton version needs to be checked.
