+++ topoptlab/elements/nonlinear_elasticity_2d.py
@@ -1,19 +1,26 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
-from typing import Any
+from typing import Any, Callable
 
 import numpy as np
 
-from topoptlab.elements.bilinear_quadrilateral import bmatrix
+from topoptlab.elements.bilinear_quadrilateral import shape_functions_dxi
+from topoptlab.elements.isoparam_mapping import invjacobian
+from topoptlab.elements.strain_measures import dispgrad_matrix,\
+                                               lagrangian_strainvar_matrix
 from topoptlab.fem import get_integrpoints
 
-def _lk_nonlinear_elast_2d(xe: np.ndarray, c: np.ndarray,
+def _lk_nonlinear_elast_2d(xe: np.ndarray,
+                           ue: np.ndarray,
+                           c: Callable,
+                           s: Callable,
                            quadr_method: str = "gauss-legendre",
                            t: np.ndarray = np.array([1.]),
                            nquad: int = 2,
                            **kwargs: Any) -> np.ndarray:
     """
-    Create element stiffness matrix for 2D linear elasticity with
-    bilinear quadrilateral Lagrangian elements.
+    Create element stiffness matrix for 2D nonlinear elasticity with
+    bilinear quadrilateral Lagrangian elements in terms of 2. Piola Kirchhoff
+    stress.
 
     Parameters
     ----------
@@ -21,8 +28,14 @@ def _lk_nonlinear_elast_2d(xe: np.ndarray, c: np.ndarray,
         coordinates of element nodes. Please look at the
         definition/function of the shape function, then the node ordering is
         clear.
-    c : np.ndarray, shape (nels,3,3) or
-        stiffness tensor.
+    ue : np.ndarray,shape (nels,8).
+        nodal displacements.
+    c : callable
+        constitutive tensor in Voigt notation (same as stiffness tensor in 
+        linear elasticity).
+    s : callable 
+        2. PK stress  tensor in Voigt notation (same as stiffness tensor in 
+        linear elasticity).
     quadr_method: str or callable
         name of quadrature method or function/callable that returns coordinates of
         quadrature points and weights. Check function get_integrpoints for
@@ -38,6 +51,7 @@ def _lk_nonlinear_elast_2d(xe: np.ndarray, c: np.ndarray,
         element stiffness matrix.
 
     """
+    ndim=2
     #
     if len(xe.shape) == 2:
         xe = xe[None,:,:]
@@ -53,15 +67,43 @@ def _lk_nonlinear_elast_2d(xe: np.ndarray, c: np.ndarray,
     nq =w.shape[0]
     #
     xi,eta = [_x[:,0] for _x in np.split(x, 2,axis=1)]
-    #
-    B,detJ = bmatrix(xi=xi, eta=eta, xe=xe,
-                     all_elems=True,
-                     return_detJ=True)
+    # calculate displacement gradient 
+    B_h,detJ = dispgrad_matrix(xi=xi, eta=eta, zeta=None, xe=xe,
+                               shape_functions_dxi=shape_functions_dxi,
+                               invjacobian=invjacobian,
+                               all_elems=True,
+                               return_detJ=True) 
+    B_h = B_h.reshape(nel, nq,  B_h.shape[-2], B_h.shape[-1])
     detJ = detJ.reshape(nel,nq)
-    B = B.reshape(nel, nq,  B.shape[-2], B.shape[-1])
+    # calculate convert to matrix form
+    F = (B_h@ue[:,None,:,None]).reshape(nel,nq,ndim,ndim) \
+         + np.eye(ndim)[None,None,:,:]
     #
-    integral = B.transpose([0,1,3,2])@c[:,None,:,:]@B
[…]
