### 2026-01-23 bb047cb
Added Gauss-Lobatto.
