+++ topoptlab/elements/huhu_2d.py
@@ -13,8 +13,8 @@ def _lk_huhu_2d(xe: np.ndarray,
                 ue: np.ndarray,
                 exponent: np.ndarray,
                 kr: np.ndarray,
-                mode="picard",
-                quadr_method: str = "gauss-legendre",
+                mode="newton",
+                quadr_method: str = "gauss-lobatto",
                 t: np.ndarray = np.array([1.]),
                 nquad: int = 4,
                 **kwargs: Any) -> np.ndarray:
@@ -132,7 +132,7 @@ if __name__ == "__main__":
     print(_lk_huhu_2d(xe = np.array([[[-1,-1],[1,-1],[1,1],[-1,1]],
                                [[-1,-1],[1,-1],[1,1],[-1,1]]]),
                 ue = np.array([[0.,0.,0.,0.,
-                                0.,-1.9999,0.,-1.9999],
+                                0.1,0.2,0.,0.],
                                [0.,0.,0.,0.,
-                                0.,-1.9999,0.,-1.9999],]),
+                                0.1,0.2,0.,0.]]),
                 exponent=6., kr=1e-6))
