### 2026-01-23 4817338
Huhu 3d implemented, but hard test still needed.
