### 2026-01-24 6b72085
Added neo-hooke and st. venant.need sensible tests.
