+++ topoptlab/material_models/neohooke.py
@@ -1,0 +1,219 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+from typing import Any,Tuple,Union
+
+import numpy as np
+
+from topoptlab.utils import to_voigt,safe_inverse,voigt_pair
+
+def eng_density(F: np.ndarray,
+                C: Union[None,np.ndarray],
+                Fdet: Union[None,np.ndarray],
+                h: np.ndarray,
+                mu: np.ndarray,
+                **kwargs: Any) -> np.ndarray:
+    """
+    Energy density of Neo-Hookean material:
+        
+        eng_dens = 1/2 (h*ln(det(F))**2 + mu *((tr(C)-3)-2*ln(det(F))))
+        
+    C is the right Cauchy–Green deformation tensor
+        
+        C = F.T @ F
+        
+    F the deformation gradient, h and mu the first and second Lame constants. 
+    det() and inv() are matrix determinant and inverse.
+    
+    Parameters
+    ----------
+    F : np.ndarray
+        deformation gradient in matrix form of shape (...,ndim,ndim). Ignored 
+        if both C and Fdet are not None.
+    C : None or np.ndarray
+        Cauchy–Green deformation tensor in matrix notation (...,ndim,ndim).
+    Fdet : np.ndarray
+        determinant of deformation gradient.
+    h : np.ndarray
+        first Lame constant.
+    mu : np.ndarray
+        second Lame constant.
+    
+    Returns
+    -------
+    eng_density : np.ndarray
+        strain energy density of shape (...).
+    """
+    #
+    if C is None:
+        #
+        C = F.swapaxes(-1,-2) @ F 
+    #
+    if Fdet is None:
+        #
+        Fdet = np.linalg.det(F)
+    #
+    return (h*np.log(Fdet)**2 + mu *((np.trace(C, axis1=-2, axis2=-1)-3)-2*np.log(Fdet)))/2
+
+def stress_2pk(F: np.ndarray,
+               Cinv: Union[None,np.ndarray],
+               C: Union[None,np.ndarray],
+               Fdet: Union[None,np.ndarray],
+               h: np.ndarray,
+               mu: np.ndarray,
+               **kwargs: Any) -> np.ndarray:
+    """
+    2. Piola Kirchhoff stress (2PK) in Voigt notation of Neo-Hookean material:
+        
+        S = h*ln(det(F)) inv(C) + mu *(I-inv(C))
+        
+    E is the Green-Lagrangian strain tensor in Voigt notation and c the 
+    stiffness tensor also in Voigt notation.
+    
+    Parameters
+    ----------
+    F : np.ndarray
+        deformation gradient in matrix form of shape (...,ndim,ndim). Ignored 
+        if both C and Fdet are not None.
+    Cinv : None or np.ndarray
+        inverse of Cauchy–Green deformation tensor in matrix notation 
+        (...,ndim,ndim).
+    C : None or np.ndarray
+        Cauchy–Green deformation tensor in matrix notation (...,ndim,ndim).
+    Fdet : np.ndarray
+        determinant of deformation gradient.
+    h : np.ndarray
+        first Lame constant.
+    mu : np.ndarray
+        second Lame constant.
+    
+    Returns
+    -------
+    s : np.ndarray
+        2. Piola Kirchhoff stress in Voigt notation (..., ndim*(ndim+1)/2).
+    """
+    #
+    if Cinv is None and C is None:
+        #
+        Cinv = safe_inverse(A=F.swapaxes(-1,-2)@F)
[…]
