### 2026-02-02 15b57b0
Update paper.md
