### 2026-02-02 #9105 @danielskatz
@stefanhiemer - as track editor, I've proofread your paper, and suggest the changes in https://github.com/stefanhiemer/topoptlab/pull/5 - please merge this, or let me know what you disagree with.
