### 2026-02-02 #9105 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
