### 2026-02-02 #9105 @danielskatz
Congratulations to @stefanhiemer (Stefan Hiemer) on your JOSS publication!!

JOSS completely depends on volunteers; we couldn't do this without you
